+++ lab-experiments/kettle-boiler/src/App.tsx
@@ -1,6 +1,6 @@
 // import { createSignal } from 'solid-js'
 import './App.css'
-import { Apparatus } from './components/Apparatus/Apparatus';
+import { Apparatus } from './components/Apparatus';
 import { HamburgerMenu } from './components/Hamburger/Hamburger';
 import worksheet from './media/testMe.txt'
