+++ lab-experiments/kettle-boiler/src/components/BeakerSystem/BeakerSystem.tsx
@@ -1,0 +1,8 @@
+import type { Component } from "solid-js";
+
+export const BeakerSystem: Component = () => {
+  return (
+    <></>
+)};
+
+export default BeakerSystem;
