+++ lab-experiments/kettle-boiler/src/components/Interactables/BallValve.tsx
@@ -1,0 +1,8 @@
+import type { Component } from "solid-js";
+
+export const BallValve: Component = () => {
+  return (
+    <></>
+)};
+
+export default BallValve;
