+++ lab-experiments/kettle-boiler/src/components/Interactables/Display.tsx
@@ -1,0 +1,8 @@
+import type { Component } from "solid-js";
+
+export const Display: Component = () => {
+  return (
+    <></>
+)};
+
+export default Display;
