+++ lab-experiments/kettle-boiler/src/components/Interactables/Gauge.tsx
@@ -1,0 +1,8 @@
+import type { Component } from "solid-js";
+
+export const Gauge: Component = () => {
+  return (
+    <></>
+)};
+
+export default Gauge;
