+++ lab-experiments/kettle-boiler/src/components/Interactables/GlobeValve.tsx
@@ -1,0 +1,8 @@
+import type { Component } from "solid-js";
+
+export const GlobeValve: Component = () => {
+  return (
+    <></>
+)};
+
+export default GlobeValve;
