+++ lab-experiments/kettle-boiler/src/components/Interactables/Rotameter.tsx
@@ -1,20 +1,7 @@
 import type { Component } from "solid-js";
-import staticElements from "./StaticElements";
-import StaticElements from "./StaticElements";
-import StaticDefs from "./StaticDefs";
-console.log(staticElements);
-export const Apparatus: Component = () => {
 
-return (<svg
-  width="1133"
-  height="777"
-  viewBox="0 0 1133 777"
-  fill="none"
-  xmlns="http://www.w3.org/2000/svg"
->
-  <g id="canvas" clip-path="url(#clip0_6_626)">
-    <rect width="1133" height="777" fill="white" />
-    <StaticElements/>
+const Rotameter: Component = () => {
+  return (
     <g id="rotameter">
       <path
         id="ball"
@@ -395,2107 +382,6 @@ return (<svg
         fill="black"
       />
     </g>
-    <g id="kettle">
-      <g id="interior">
-        <path
-          id="Rectangle 72"
-          d="M511.5 392.5V402C511.5 403.381 510.381 404.5 509 404.5H487C485.619 404.5 484.5 403.381 484.5 402V392.5H511.5Z"
-          fill="url(#paint44_linear_6_626)"
-          stroke="black"
-        />
-        <path
-          id="Rectangle 74_2"
-          d="M1025.5 214.5V209C1025.5 208.172 1024.83 207.5 1024 207.5H1000C999.172 207.5 998.5 208.172 998.5 209V214.5H1025.5Z"
-          fill="url(#paint45_linear_6_626)"
-          stroke="black"
-        />
-        <path
-          id="Rectangle 75_2"
-          d="M854.5 214.5V209C854.5 208.172 853.828 207.5 853 207.5H829C828.172 207.5 827.5 208.172 827.5 209V214.5H854.5Z"
-          fill="url(#paint46_linear_6_626)"
-          stroke="black"
-        />
-        <path
-          id="Rectangle 73_2"
-          d="M978.5 397.5V402C978.5 403.381 977.381 404.5 976 404.5H954C952.619 404.5 951.5 403.381 951.5 402V397.5H978.5Z"
-          fill="url(#paint47_linear_6_626)"
-          stroke="black"
-        />
-        <path
-          id="walls"
-          d="M1018 214.011H1023.48C1023.48 214.011 1085 211.05 1085 306.011C1085 400.972 1023.48 397.936 1023.48 397.936H1018V398H529.214C519.75 398 510.298 397.329 500.93 395.99L494.07 395.011C484.702 393.672 475.25 393 465.786 393H435V273C440.816 273 446.449 270.965 450.922 267.248L498.08 228.061C508.758 219.188 522.139 214.238 536 214.011V214H1018V214.011Z"
-          fill="url(#paint48_linear_6_626)"
-        />
-        <g id="trap">
-          <rect
-            id="Rectangle 63"
-            x="435.5"
-            y="283.5"
-            width="430"
-            height="19"
-            fill="url(#paint49_linear_6_626)"
-            stroke="black"
-          />
-          <rect
-            id="Rectangle 65"
-            x="435.5"
-            y="307.5"
-            width="430"
-            height="19"
-            fill="url(#paint50_linear_6_626)"
-            stroke="black"
-          />
-          <rect
-            id="Rectangle 66"
-            x="435.5"
-            y="339.5"
-            width="430"
-            height="19"
-            fill="url(#paint51_linear_6_626)"
-            stroke="black"
-          />
-          <rect
-            id="Rectangle 64"
-            x="435.5"
-            y="363.5"
-            width="430"
-            height="19"
-            fill="url(#paint52_linear_6_626)"
-            stroke="black"
-          />
-          <g id="Subtract">
-            <mask id="path-161-inside-3_6_626" fill="white">
-              <path
-                d="M865.538 277C889.132 282.167 907 306.614 907 335.994C907 365.373 889.132 389.819 865.538 394.986V277Z"
-              />
-            </mask>
-            <path
-              d="M865.538 277C889.132 282.167 907 306.614 907 335.994C907 365.373 889.132 389.819 865.538 394.986V277Z"
-              fill="#686155"
-            />
-            <path
-              d="M865.538 277L865.752 276.024L864.538 275.758V277H865.538ZM865.538 394.986H864.538V396.229L865.752 395.963L865.538 394.986ZM865.538 277L865.324 277.977C888.363 283.022 906 306.978 906 335.994H907H908C908 306.25 889.901 281.311 865.752 276.024L865.538 277ZM907 335.994H906C906 365.009 888.363 388.964 865.324 394.009L865.538 394.986L865.752 395.963C889.9 390.675 908 365.738 908 335.994H907ZM865.538 394.986H866.538V277H865.538H864.538V394.986H865.538Z"
-              fill="black"
-              mask="url(#path-161-inside-3_6_626)"
-            />
-          </g>
-          <rect
-            id="Rectangle 57"
-            x="755.5"
-            y="279.5"
-            width="4"
-            height="9"
-            rx="1.5"
-            fill="url(#paint53_linear_6_626)"
-            stroke="black"
-          />
-          <rect
-            id="Rectangle 67"
-            x="436.5"
-            y="281.5"
-            width="318"
-            height="5"
-            fill="url(#paint54_linear_6_626)"
-            stroke="black"
-          />
-          <rect
-            id="Rectangle 68"
-            x="755.5"
-            y="377.5"
-            width="4"
-            height="9"
-            rx="1.5"
-            fill="url(#paint55_linear_6_626)"
-            stroke="black"
-          />
-          <rect
-            id="Rectangle 69"
-            x="436.5"
-            y="379.5"
-            width="318"
-            height="5"
-            fill="url(#paint56_linear_6_626)"
-            stroke="black"
-          />
-          <g id="Repeat group 1_2">
-            <rect
-              id="Rectangle 62"
-              x="533.5"
-              y="277.5"
-              width="3"
-              height="117"
-              rx="0.5"
-              fill="#686155"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 62_2"
-              x="643.5"
-              y="277.5"
-              width="3"
-              height="117"
-              rx="0.5"
-              fill="#686155"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 62_3"
-              x="753.5"
-              y="277.5"
-              width="3"
-              height="117"
-              rx="0.5"
-              fill="#686155"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 62_4"
-              x="863.5"
-              y="277.5"
-              width="3"
-              height="117"
-              rx="0.5"
-              fill="#686155"
-              stroke="black"
-            />
-          </g>
-        </g>
-        <path
-          id="chamberFill"
-          d="M914 398H529.214C519.75 398 510.298 397.329 500.93 395.991L494.07 395.011C484.702 393.673 475.25 393 465.786 393H438V272.817C440.34 272.533 442.62 271.92 444.775 271H914V398Z"
-          fill="#3B8CCF"
-          fill-opacity="0.6"
-        />
-        <path
-          id="overflowFill"
-          d="M1056.33 386C1040.3 398.756 1023.51 397.937 1023.48 397.936H1018V398H914V386H1056.33Z"
-          fill="#78AEDD"
-        />
-        <path
-          id="overflowPath"
-          d="M914 272.5C917 272.5 917.5 274 917.5 277V397"
-          stroke="#78AEDD"
-          stroke-width="3"
-        />
-        <rect
-          id="weir"
-          x="912.5"
-          y="274.5"
-          width="3"
-          height="123"
-          rx="0.5"
-          fill="#878E95"
-          stroke="black"
-        />
-        <path
-          id="outline"
-          d="M1017.5 214.5V214.511H1023.5L1023.51 214.51H1023.52C1023.52 214.509 1023.53 214.509 1023.55 214.509C1023.57 214.508 1023.62 214.506 1023.67 214.505C1023.78 214.502 1023.95 214.5 1024.17 214.501C1024.61 214.502 1025.27 214.517 1026.1 214.568C1027.76 214.672 1030.14 214.924 1032.99 215.521C1038.71 216.717 1046.33 219.291 1053.95 224.806C1069.16 235.818 1084.5 258.646 1084.5 306.011L1084.49 308.213C1084.02 354.085 1068.93 376.356 1053.95 387.178C1046.33 392.683 1038.71 395.248 1032.99 396.437C1030.14 397.03 1027.76 397.281 1026.1 397.382C1025.27 397.432 1024.61 397.445 1024.17 397.446C1023.95 397.447 1023.78 397.444 1023.67 397.441C1023.62 397.44 1023.57 397.439 1023.55 397.438C1023.53 397.437 1023.52 397.437 1023.52 397.437H1023.51L1023.5 397.436H1017.5V397.5H529.214C519.774 397.5 510.345 396.83 501 395.495L494.141 394.516C484.749 393.174 475.273 392.5 465.786 392.5H435.5V273.494C441.255 273.381 446.807 271.318 451.241 267.633L498.399 228.445C508.989 219.645 522.26 214.736 536.008 214.511L536.5 214.503V214.5H1017.5Z"
-          stroke="black"
-        />
-      </g>
-      <g id="exterior">
-        <path
-          id="Rectangle 58"
-          d="M536.5 397.5H529.214C519.774 397.5 510.345 396.831 501 395.496L494.142 394.515C484.75 393.173 475.274 392.5 465.786 392.5H435.5V273.495C441.255 273.381 446.807 271.318 451.241 267.633L498.399 228.446C509.116 219.541 522.577 214.621 536.5 214.505V397.5Z"
-          fill="url(#paint57_linear_6_626)"
-          stroke="black"
-        />
-        <path
-          id="Rectangle 59"
-          d="M1024.17 214.5C1024.61 214.502 1025.27 214.516 1026.1 214.568C1027.76 214.671 1030.14 214.923 1032.99 215.521C1038.71 216.716 1046.33 219.29 1053.95 224.805C1069.16 235.817 1084.5 258.646 1084.5 306.01C1084.5 353.375 1069.17 376.185 1053.95 387.178C1046.33 392.684 1038.71 395.248 1032.99 396.436C1030.14 397.03 1027.76 397.28 1026.1 397.381C1025.27 397.432 1024.62 397.445 1024.17 397.446C1023.95 397.446 1023.78 397.444 1023.67 397.441C1023.62 397.44 1023.57 397.438 1023.55 397.437C1023.53 397.437 1023.52 397.436 1023.52 397.436H1023.51L1023.5 397.435H1017V214.51H1023.51C1023.51 214.51 1023.51 214.509 1023.52 214.509C1023.52 214.509 1023.53 214.509 1023.55 214.508C1023.57 214.507 1023.62 214.507 1023.67 214.505C1023.78 214.503 1023.95 214.5 1024.17 214.5Z"
-          fill="url(#paint58_linear_6_626)"
-          stroke="black"
-        />
-        <g id="Rectangle 60">
-          <mask id="path-178-inside-4_6_626" fill="white">
-            <path d="M536 214H1018V398H536V214Z" />
-          </mask>
-          <path
-            d="M536 214H1018V398H536V214Z"
-            fill="url(#paint59_linear_6_626)"
-          />
-          <path
-            d="M536 214V215H1018V214V213H536V214ZM1018 398V397H536V398V399H1018V398Z"
-            fill="black"
-            mask="url(#path-178-inside-4_6_626)"
-          />
-        </g>
-        <g id="Subtract_2">
-          <mask id="path-180-inside-5_6_626" fill="white">
-            <path
-              d="M777 246C810.137 246 837 272.863 837 306C837 339.138 810.137 366 777 366C743.863 366 717 339.138 717 306C717 272.863 743.863 246 777 246ZM777 266C754.909 266 737 283.909 737 306C737 328.092 754.909 346 777 346C799.091 346 817 328.092 817 306C817 283.909 799.091 266 777 266Z"
-            />
-          </mask>
-          <path
-            d="M777 246C810.137 246 837 272.863 837 306C837 339.138 810.137 366 777 366C743.863 366 717 339.138 717 306C717 272.863 743.863 246 777 246ZM777 266C754.909 266 737 283.909 737 306C737 328.092 754.909 346 777 346C799.091 346 817 328.092 817 306C817 283.909 799.091 266 777 266Z"
-            fill="#83878B"
-          />
-          <path
-            d="M777 246V247C809.585 247 836 273.416 836 306H837H838C838 272.311 810.689 245 777 245V246ZM837 306H836C836 338.585 809.585 365 777 365V366V367C810.689 367 838 339.69 838 306H837ZM777 366V365C744.415 365 718 338.585 718 306H717H716C716 339.69 743.311 367 777 367V366ZM717 306H718C718 273.416 744.415 247 777 247V246V245C743.311 245 716 272.311 716 306H717ZM777 266V265C754.356 265 736 283.357 736 306H737H738C738 284.461 755.461 267 777 267V266ZM737 306H736C736 328.644 754.356 347 777 347V346V345C755.461 345 738 327.54 738 306H737ZM777 346V347C799.644 347 818 328.644 818 306H817H816C816 327.54 798.539 345 777 345V346ZM817 306H818C818 283.357 799.644 265 777 265V266V267C798.539 267 816 284.461 816 306H817Z"
-            fill="black"
-            mask="url(#path-180-inside-5_6_626)"
-          />
-        </g>
-        <g id="Repeat group 1_3">
-          <g id="Ellipse 2">
-            <circle
-              cx="757.958"
-              cy="260.17"
-              r="3.5"
-              transform="rotate(-22.5 757.958 260.17)"
-              fill="#C3C5C8"
-            />
-            <circle
-              cx="757.958"
-              cy="260.17"
-              r="3.5"
-              transform="rotate(-22.5 757.958 260.17)"
-              stroke="black"
-            />
-            <circle
-              cx="757.958"
-              cy="260.17"
-              r="3.5"
-              transform="rotate(-22.5 757.958 260.17)"
-              stroke="black"
-              stroke-opacity="0.2"
-            />
-          </g>
-          <g id="Ellipse 2_2">
-            <circle
-              cx="795.844"
-              cy="260.169"
-              r="3.5"
-              transform="rotate(22.5 795.844 260.169)"
-              fill="#C3C5C8"
-            />
-            <circle
-              cx="795.844"
-              cy="260.169"
-              r="3.5"
-              transform="rotate(22.5 795.844 260.169)"
-              stroke="black"
-            />
-            <circle
-              cx="795.844"
-              cy="260.169"
-              r="3.5"
-              transform="rotate(22.5 795.844 260.169)"
-              stroke="black"
-              stroke-opacity="0.2"
-            />
-          </g>
-          <g id="Ellipse 2_3">
-            <circle
-              cx="822.633"
-              cy="286.959"
-              r="3.5"
-              transform="rotate(67.5 822.633 286.959)"
-              fill="#C3C5C8"
-            />
-            <circle
-              cx="822.633"
-              cy="286.959"
-              r="3.5"
-              transform="rotate(67.5 822.633 286.959)"
-              stroke="black"
-            />
-            <circle
-              cx="822.633"
-              cy="286.959"
-              r="3.5"
-              transform="rotate(67.5 822.633 286.959)"
-              stroke="black"
-              stroke-opacity="0.2"
-            />
-          </g>
-          <g id="Ellipse 2_4">
-            <circle
-              cx="822.633"
-              cy="324.844"
-              r="3.5"
-              transform="rotate(112.5 822.633 324.844)"
-              fill="#C3C5C8"
-            />
-            <circle
-              cx="822.633"
-              cy="324.844"
-              r="3.5"
-              transform="rotate(112.5 822.633 324.844)"
-              stroke="black"
-            />
-            <circle
-              cx="822.633"
-              cy="324.844"
-              r="3.5"
-              transform="rotate(112.5 822.633 324.844)"
-              stroke="black"
-              stroke-opacity="0.2"
-            />
-          </g>
-          <g id="Ellipse 2_5">
-            <circle
-              cx="795.844"
-              cy="351.634"
-              r="3.5"
-              transform="rotate(157.5 795.844 351.634)"
-              fill="#C3C5C8"
-            />
-            <circle
-              cx="795.844"
-              cy="351.634"
-              r="3.5"
-              transform="rotate(157.5 795.844 351.634)"
-              stroke="black"
-            />
-            <circle
-              cx="795.844"
-              cy="351.634"
-              r="3.5"
-              transform="rotate(157.5 795.844 351.634)"
-              stroke="black"
-              stroke-opacity="0.2"
-            />
-          </g>
-          <g id="Ellipse 2_6">
-            <circle
-              cx="757.958"
-              cy="351.634"
-              r="3.5"
-              transform="rotate(-157.5 757.958 351.634)"
-              fill="#C3C5C8"
-            />
-            <circle
-              cx="757.958"
-              cy="351.634"
-              r="3.5"
-              transform="rotate(-157.5 757.958 351.634)"
-              stroke="black"
-            />
-            <circle
-              cx="757.958"
-              cy="351.634"
-              r="3.5"
-              transform="rotate(-157.5 757.958 351.634)"
-              stroke="black"
-              stroke-opacity="0.2"
-            />
-          </g>
-          <g id="Ellipse 2_7">
-            <circle
-              cx="731.169"
-              cy="324.844"
-              r="3.5"
-              transform="rotate(-112.5 731.169 324.844)"
-              fill="#C3C5C8"
-            />
-            <circle
-              cx="731.169"
-              cy="324.844"
-              r="3.5"
-              transform="rotate(-112.5 731.169 324.844)"
-              stroke="black"
-            />
-            <circle
-              cx="731.169"
-              cy="324.844"
-              r="3.5"
-              transform="rotate(-112.5 731.169 324.844)"
-              stroke="black"
-              stroke-opacity="0.2"
-            />
-          </g>
-          <g id="Ellipse 2_8">
-            <circle
-              cx="731.169"
-              cy="286.959"
-              r="3.5"
-              transform="rotate(-67.5 731.169 286.959)"
-              fill="#C3C5C8"
-            />
-            <circle
-              cx="731.169"
-              cy="286.959"
-              r="3.5"
-              transform="rotate(-67.5 731.169 286.959)"
-              stroke="black"
-            />
-            <circle
-              cx="731.169"
-              cy="286.959"
-              r="3.5"
-              transform="rotate(-67.5 731.169 286.959)"
-              stroke="black"
-              stroke-opacity="0.2"
-            />
-          </g>
-        </g>
-        <circle id="Ellipse 3" cx="777" cy="306" r="40" fill="#161515" />
-      </g>
-      <g id="base">
-        <rect
-          id="Rectangle 70"
-          x="367.5"
-          y="261.5"
-          width="27"
-          height="14"
-          rx="2.5"
-          fill="url(#paint60_linear_6_626)"
-          stroke="black"
-        />
-        <rect
-          id="Rectangle 71"
-          x="367.5"
-          y="390.5"
-          width="27"
-          height="14"
-          rx="2.5"
-          fill="url(#paint61_linear_6_626)"
-          stroke="black"
-        />
-        <rect
-          id="Rectangle 53"
-          x="341.5"
-          y="273.5"
-          width="79"
-          height="119"
-          fill="url(#paint62_linear_6_626)"
-          stroke="black"
-        />
-        <g id="Group 1">
-          <g id="Repeat group 1_4">
-            <rect
-              id="Rectangle 56"
-              x="414.5"
-              y="268.5"
-              width="27"
-              height="9"
-              rx="2.5"
-              fill="url(#paint63_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_2"
-              x="414.5"
-              y="288.5"
-              width="27"
-              height="9"
-              rx="2.5"
-              fill="url(#paint64_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_3"
-              x="414.5"
-              y="308.5"
-              width="27"
-              height="9"
-              rx="2.5"
-              fill="url(#paint65_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_4"
-              x="414.5"
-              y="328.5"
-              width="27"
-              height="9"
-              rx="2.5"
-              fill="url(#paint66_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_5"
-              x="414.5"
-              y="348.5"
-              width="27"
-              height="9"
-              rx="2.5"
-              fill="url(#paint67_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_6"
-              x="414.5"
-              y="368.5"
-              width="27"
-              height="9"
-              rx="2.5"
-              fill="url(#paint68_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_7"
-              x="414.5"
-              y="388.5"
-              width="27"
-              height="9"
-              rx="2.5"
-              fill="url(#paint69_linear_6_626)"
-              stroke="black"
-            />
-          </g>
-          <rect
-            id="Rectangle 54"
-            x="418.5"
-            y="263.5"
-            width="19"
-            height="139"
-            rx="1.5"
-            fill="url(#paint70_linear_6_626)"
-            stroke="black"
-          />
-        </g>
-        <g id="Group 2">
-          <g id="Repeat group 2_2">
-            <rect
-              id="Rectangle 56_8"
-              x="324.5"
-              y="268.5"
-              width="23"
-              height="9"
-              rx="2.5"
-              fill="url(#paint71_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_9"
-              x="324.5"
-              y="288.5"
-              width="23"
-              height="9"
-              rx="2.5"
-              fill="url(#paint72_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_10"
-              x="324.5"
-              y="308.5"
-              width="23"
-              height="9"
-              rx="2.5"
-              fill="url(#paint73_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_11"
-              x="324.5"
-              y="328.5"
-              width="23"
-              height="9"
-              rx="2.5"
-              fill="url(#paint74_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_12"
-              x="324.5"
-              y="348.5"
-              width="23"
-              height="9"
-              rx="2.5"
-              fill="url(#paint75_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_13"
-              x="324.5"
-              y="368.5"
-              width="23"
-              height="9"
-              rx="2.5"
-              fill="url(#paint76_linear_6_626)"
-              stroke="black"
-            />
-            <rect
-              id="Rectangle 56_14"
-              x="324.5"
-              y="388.5"
-              width="23"
-              height="9"
-              rx="2.5"
-              fill="url(#paint77_linear_6_626)"
-              stroke="black"
-            />
-          </g>
-          <rect
-            id="Rectangle 55"
-            x="328.5"
-            y="263.5"
-            width="15"
-            height="139"
-            rx="1.5"
-            fill="url(#paint78_linear_6_626)"
-            stroke="black"
-          />
-          <line
-            id="Line 4"
-            x1="336"
-            y1="264"
-            x2="336"
-            y2="402"
-            stroke="#363636"
-            stroke-width="2"
-          />
-        </g>
-        <rect
-          id="Rectangle 57_2"
-          x="425.5"
-          y="263.5"
-          width="5"
-          height="139"
-          fill="#43392A"
-          stroke="black"
-        />
-      </g>
-    </g>
-    <g id="display">
-      <rect
-        id="Rectangle 74_3"
-        x="347.5"
-        y="565.5"
-        width="66"
-        height="33"
-        rx="3.5"
-        fill="#888888"
-        stroke="black"
-      />
-      <rect
-        id="condTempScreen"
-        x="350.5"
-        y="568.5"
-        width="60"
-        height="27"
-        rx="3.5"
-        fill="#111111"
-        stroke="black"
-      />
-    </g>
-    <g id="display_2">
-      <rect
-        id="Rectangle 74_4"
-        x="576.5"
-        y="428.5"
-        width="66"
-        height="33"
-        rx="3.5"
-        fill="#888888"
-        stroke="black"
-      />
-      <rect
-        id="feedTempScreen"
-        x="579.5"
-        y="431.5"
-        width="60"
-        height="27"
-        rx="3.5"
-        fill="#111111"
-        stroke="black"
-      />
-    </g>
-    <g id="display_3">
-      <rect
-        id="Rectangle 74_5"
-        x="220.5"
-        y="351.5"
-        width="66"
-        height="33"
-        rx="3.5"
-        fill="#888888"
-        stroke="black"
-      />
-      <rect
-        id="steamTempScreen"
-        x="223.5"
-        y="354.5"
-        width="60"
-        height="27"
-        rx="3.5"
-        fill="#111111"
-        stroke="black"
-      />
-    </g>
-    <g id="display_4">
-      <rect
-        id="Rectangle 74_6"
-        x="820.5"
-        y="479.5"
-        width="66"
-        height="33"
-        rx="3.5"
-        fill="#888888"
-        stroke="black"
-      />
-      <rect
-        id="outTempScreen"
-        x="823.5"
-        y="482.5"
-        width="60"
-        height="27"
-        rx="3.5"
-        fill="#111111"
-        stroke="black"
-      />
-    </g>
-    <path
-      id="condFall"
-      d="M167 522C160.333 522 153.667 522 147 522C147.667 524.57 148.741 527.141 149 529.711C149.013 529.841 149.026 529.97 149.039 530.1C149.657 536.197 150.376 542.294 150.5 548.391C151.364 590.894 152.077 633.397 152.852 675.9C152.901 678.6 152.95 681.3 153 684C155.667 684 158.333 684 161 684C161.05 681.3 161.099 678.6 161.148 675.9C161.923 633.397 162.636 590.894 163.5 548.391C163.624 542.294 164.343 536.197 164.961 530.1C164.974 529.97 164.987 529.841 165 529.711C165.259 527.141 166.333 524.57 167 522Z"
-      fill="#3B8CCF"
-      fill-opacity="0.6"
-    />
-    <path
-      id="condFall_2"
-      d="M975 522C968.333 522 961.667 522 955 522C955.667 524.57 956.741 527.141 957 529.711C957.013 529.841 957.026 529.97 957.039 530.1C957.657 536.197 958.376 542.294 958.5 548.391C959.364 590.894 960.077 633.397 960.852 675.9C960.901 678.6 960.95 681.3 961 684C963.667 684 966.333 684 969 684C969.05 681.3 969.099 678.6 969.148 675.9C969.923 633.397 970.636 590.894 971.5 548.391C971.624 542.294 972.343 536.197 972.961 530.1C972.974 529.97 972.987 529.841 973 529.711C973.259 527.141 974.333 524.57 975 522Z"
-      fill="#3B8CCF"
-      fill-opacity="0.6"
-    />
-    <g id="gauge">
-      <rect
-        id="Rectangle 53_2"
-        x="-0.5"
-        y="-0.5"
-        width="9.03082"
-        height="15.1003"
-        rx="0.5"
-        transform="matrix(0 -1 -1 0 388.05 104)"
-        fill="url(#paint79_linear_6_626)"
-        stroke="black"
-      />
-      <path
-        id="Ellipse 5"
-        d="M381 35.2842C398.509 35.2842 412.701 49.4329 412.701 66.8829C412.701 84.3328 398.509 98.4815 381 98.4815C363.491 98.4813 349.299 84.3327 349.299 66.8829C349.299 49.433 363.491 35.2844 381 35.2842Z"
-        fill="#5B8BF3"
-        stroke="black"
-      />
-      <path
-        id="Ellipse 6"
-        d="M381 37.2904C397.398 37.2906 410.688 50.5414 410.688 66.8832C410.688 83.2249 397.398 96.4747 381 96.475C364.603 96.475 351.312 83.225 351.312 66.8832C351.312 50.5412 364.603 37.2904 381 37.2904Z"
-        fill="#D9D9D9"
-        stroke="black"
-      />
-      <g id="minor">
-        <path
-          id="Line 14"
-          d="M385.093 41.1238L384.778 43.1053"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_2"
-          d="M389.085 42.0792L388.463 43.9872"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_3"
-          d="M392.878 43.6453L391.964 45.4328"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_4"
-          d="M396.378 45.7836L395.195 47.4067"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_5"
-          d="M402.166 51.5533L400.538 52.7325"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_6"
-          d="M404.311 55.0428L402.518 55.9536"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_7"
-          d="M405.882 58.8237L403.969 59.4434"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_8"
-          d="M406.841 62.803L404.853 63.1168"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_9"
-          d="M406.841 70.9627L404.853 70.6489"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_10"
-          d="M405.882 74.9421L403.969 74.3223"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_11"
-          d="M404.311 78.723L402.518 77.8122"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_12"
-          d="M402.166 82.2124L400.538 81.0332"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_13"
-          d="M359.834 82.2124L361.462 81.0332"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_14"
-          d="M357.689 78.723L359.482 77.8122"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_15"
-          d="M356.118 74.9421L358.031 74.3223"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_16"
-          d="M355.159 70.9627L357.147 70.6489"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_17"
-          d="M355.159 62.803L357.147 63.1168"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_18"
-          d="M356.118 58.8237L358.031 59.4434"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_19"
-          d="M357.689 55.0427L359.482 55.9535"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_20"
-          d="M359.834 51.5533L361.462 52.7325"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_21"
-          d="M365.622 45.7836L366.805 47.4067"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_22"
-          d="M369.122 43.6453L370.036 45.4328"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_23"
-          d="M372.915 42.0792L373.537 43.9872"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 14_24"
-          d="M376.907 41.1238L377.222 43.1053"
-          stroke="black"
-          stroke-width="0.5"
-        />
-      </g>
-      <g id="major">
-        <path
-          id="Line 15"
-          d="M381 40.8027V44.8151"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 15_2"
-          d="M399.5 48.4414L396.654 51.2786"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 15_3"
-          d="M407.163 66.8829L403.138 66.8829"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 15_4"
-          d="M399.5 85.3243L396.654 82.4871"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 15_5"
-          d="M362.5 85.3243L365.346 82.4871"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 15_6"
-          d="M354.837 66.8829L358.862 66.8829"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Line 15_7"
-          d="M362.5 48.4414L365.346 51.2786"
-          stroke="black"
-          stroke-width="0.5"
-        />
-      </g>
-      <path
-        id="Ellipse 8"
-        d="M399.5 85.3243C409.717 75.1394 409.717 58.6264 399.5 48.4414C389.283 38.2565 372.717 38.2565 362.5 48.4414C352.283 58.6264 352.283 75.1394 362.5 85.3243"
-        stroke="black"
-        stroke-width="0.5"
-        stroke-linecap="square"
-      />
-      <path
-        id="psig"
-        d="M375.423 91.1626V86.6626H375.908V87.1825H375.968C376.005 87.1256 376.056 87.0532 376.121 86.9651C376.188 86.8756 376.283 86.7961 376.407 86.7265C376.532 86.6555 376.701 86.62 376.914 86.62C377.19 86.62 377.433 86.6889 377.643 86.8266C377.853 86.9644 378.017 87.1597 378.135 87.4126C378.253 87.6654 378.312 87.9637 378.312 88.3075C378.312 88.6541 378.253 88.9545 378.135 89.2087C378.017 89.4616 377.854 89.6576 377.645 89.7968C377.436 89.9346 377.195 90.0035 376.923 90.0035C376.712 90.0035 376.544 89.9687 376.418 89.8991C376.291 89.8281 376.194 89.7478 376.126 89.6583C376.058 89.5674 376.005 89.4921 375.968 89.4325H375.925V91.1626H375.423ZM375.917 88.2989C375.917 88.5461 375.953 88.7641 376.026 88.9531C376.098 89.1406 376.204 89.2876 376.343 89.3941C376.482 89.4992 376.653 89.5518 376.854 89.5518C377.065 89.5518 377.24 89.4964 377.381 89.3856C377.523 89.2734 377.629 89.1228 377.7 88.9339C377.773 88.7435 377.809 88.5319 377.809 88.2989C377.809 88.0688 377.773 87.8614 377.702 87.6768C377.633 87.4907 377.527 87.3437 377.385 87.2357C377.244 87.1264 377.068 87.0717 376.854 87.0717C376.65 87.0717 376.478 87.1235 376.339 87.2272C376.2 87.3295 376.095 87.4729 376.023 87.6576C375.952 87.8408 375.917 88.0546 375.917 88.2989ZM381.397 87.3955L380.945 87.5234C380.917 87.4481 380.875 87.3749 380.82 87.3039C380.766 87.2315 380.692 87.1718 380.598 87.1249C380.504 87.0781 380.384 87.0546 380.238 87.0546C380.038 87.0546 379.871 87.1008 379.737 87.1931C379.605 87.284 379.539 87.3998 379.539 87.5404C379.539 87.6654 379.585 87.7641 379.675 87.8366C379.766 87.909 379.908 87.9694 380.102 88.0177L380.587 88.137C380.88 88.208 381.098 88.3167 381.242 88.463C381.385 88.6079 381.457 88.7947 381.457 89.0234C381.457 89.2109 381.403 89.3785 381.295 89.5262C381.188 89.6739 381.039 89.7904 380.847 89.8756C380.656 89.9609 380.433 90.0035 380.178 90.0035C379.845 90.0035 379.568 89.931 379.349 89.7862C379.131 89.6413 378.992 89.4296 378.934 89.1512L379.411 89.0319C379.457 89.208 379.543 89.3401 379.669 89.4282C379.797 89.5163 379.964 89.5603 380.17 89.5603C380.404 89.5603 380.59 89.5106 380.728 89.4112C380.867 89.3103 380.937 89.1896 380.937 89.0489C380.937 88.9353 380.897 88.8401 380.818 88.7634C380.738 88.6853 380.616 88.6271 380.451 88.5887L379.906 88.4609C379.606 88.3898 379.386 88.2798 379.245 88.1306C379.106 87.98 379.036 87.7918 379.036 87.566C379.036 87.3813 379.088 87.218 379.192 87.0759C379.297 86.9339 379.44 86.8224 379.62 86.7414C379.802 86.6604 380.008 86.62 380.238 86.62C380.562 86.62 380.816 86.691 381.001 86.833C381.187 86.9751 381.319 87.1626 381.397 87.3955ZM382.214 89.9353V86.6626H382.716V89.9353H382.214ZM382.469 86.1171C382.371 86.1171 382.287 86.0837 382.216 86.017C382.146 85.9502 382.111 85.87 382.111 85.7762C382.111 85.6825 382.146 85.6022 382.216 85.5354C382.287 85.4687 382.371 85.4353 382.469 85.4353C382.567 85.4353 382.651 85.4687 382.721 85.5354C382.792 85.6022 382.827 85.6825 382.827 85.7762C382.827 85.87 382.792 85.9502 382.721 86.017C382.651 86.0837 382.567 86.1171 382.469 86.1171ZM384.958 91.2308C384.716 91.2308 384.507 91.1995 384.332 91.137C384.157 91.0759 384.012 90.995 383.895 90.8941C383.78 90.7947 383.689 90.6881 383.62 90.5745L384.021 90.2933C384.066 90.3529 384.124 90.4211 384.194 90.4978C384.263 90.5759 384.358 90.6434 384.479 90.7002C384.601 90.7585 384.761 90.7876 384.958 90.7876C385.223 90.7876 385.441 90.7237 385.613 90.5958C385.784 90.468 385.87 90.2677 385.87 89.995V89.3302H385.828C385.791 89.3898 385.738 89.4637 385.67 89.5518C385.603 89.6384 385.507 89.7158 385.38 89.784C385.255 89.8508 385.086 89.8842 384.873 89.8842C384.609 89.8842 384.372 89.8217 384.162 89.6967C383.953 89.5717 383.787 89.3898 383.665 89.1512C383.544 88.9126 383.484 88.6228 383.484 88.2819C383.484 87.9467 383.543 87.6548 383.661 87.4062C383.779 87.1562 383.943 86.963 384.153 86.8266C384.363 86.6889 384.606 86.62 384.882 86.62C385.095 86.62 385.264 86.6555 385.389 86.7265C385.515 86.7961 385.612 86.8756 385.679 86.9651C385.747 87.0532 385.799 87.1256 385.836 87.1825H385.887V86.6626H386.373V90.0291C386.373 90.3103 386.309 90.539 386.182 90.7151C386.055 90.8927 385.885 91.0227 385.67 91.105C385.457 91.1889 385.22 91.2308 384.958 91.2308ZM384.941 89.4325C385.143 89.4325 385.314 89.3863 385.453 89.294C385.592 89.2016 385.698 89.0688 385.77 88.8955C385.843 88.7222 385.879 88.5148 385.879 88.2734C385.879 88.0376 385.843 87.8295 385.772 87.6491C385.701 87.4687 385.596 87.3273 385.457 87.2251C385.318 87.1228 385.146 87.0717 384.941 87.0717C384.728 87.0717 384.551 87.1256 384.409 87.2336C384.268 87.3416 384.162 87.4864 384.091 87.6683C384.022 87.8501 383.987 88.0518 383.987 88.2734C383.987 88.5006 384.022 88.7016 384.093 88.8764C384.166 89.0496 384.272 89.186 384.413 89.2854C384.555 89.3835 384.731 89.4325 384.941 89.4325Z"
-        fill="black"
-      />
-      <path
-        id="0_2"
-        d="M367.781 81.9703C367.46 81.9703 367.186 81.8829 366.961 81.7082C366.735 81.5321 366.562 81.2771 366.443 80.9433C366.324 80.6081 366.264 80.2033 366.264 79.7288C366.264 79.2572 366.324 78.8545 366.443 78.5207C366.564 78.1855 366.737 77.9298 366.963 77.7537C367.19 77.5761 367.463 77.4873 367.781 77.4873C368.099 77.4873 368.371 77.5761 368.597 77.7537C368.824 77.9298 368.998 78.1855 369.117 78.5207C369.238 78.8545 369.298 79.2572 369.298 79.7288C369.298 80.2033 369.238 80.6081 369.119 80.9433C369 81.2771 368.827 81.5321 368.601 81.7082C368.375 81.8829 368.102 81.9703 367.781 81.9703ZM367.781 81.5016C368.099 81.5016 368.346 81.3481 368.522 81.0413C368.699 80.7345 368.787 80.297 368.787 79.7288C368.787 79.351 368.746 79.0293 368.665 78.7636C368.586 78.498 368.471 78.2956 368.32 78.1564C368.171 78.0172 367.991 77.9476 367.781 77.9476C367.466 77.9476 367.219 78.1031 367.042 78.4142C366.864 78.7239 366.775 79.1621 366.775 79.7288C366.775 80.1067 366.815 80.4277 366.895 80.6919C366.974 80.9561 367.088 81.1571 367.238 81.2949C367.388 81.4327 367.569 81.5016 367.781 81.5016Z"
-        fill="black"
-      />
-      <path
-        id="5_2"
-        d="M361.684 68.9302C361.434 68.9302 361.209 68.8805 361.008 68.7811C360.808 68.6816 360.647 68.5453 360.527 68.372C360.406 68.1987 360.34 68.0012 360.329 67.7796H360.84C360.86 67.9771 360.949 68.1404 361.108 68.2697C361.269 68.3975 361.461 68.4615 361.684 68.4615C361.863 68.4615 362.022 68.4196 362.161 68.3357C362.302 68.2519 362.412 68.1369 362.491 67.9906C362.572 67.8428 362.613 67.6759 362.613 67.4899C362.613 67.2995 362.571 67.1298 362.487 66.9806C362.405 66.8301 362.291 66.7115 362.146 66.6248C362.001 66.5382 361.836 66.4941 361.65 66.4927C361.516 66.4913 361.379 66.5119 361.238 66.5545C361.098 66.5957 360.982 66.649 360.891 66.7143L360.397 66.6546L360.661 64.5069H362.928V64.9757H361.104L360.951 66.2626H360.976C361.066 66.1916 361.178 66.1326 361.313 66.0857C361.448 66.0389 361.589 66.0154 361.735 66.0154C362.002 66.0154 362.24 66.0794 362.449 66.2072C362.659 66.3336 362.824 66.5069 362.943 66.7271C363.064 66.9472 363.124 67.1987 363.124 67.4813C363.124 67.7597 363.062 68.0083 362.937 68.2271C362.813 68.4444 362.642 68.6163 362.425 68.7427C362.208 68.8677 361.961 68.9302 361.684 68.9302Z"
-        fill="black"
-      />
-      <path
-        id="10_2"
-        d="M366.886 51.4669V55.8305H366.357V52.0208H366.332L365.266 52.7282V52.1913L366.357 51.4669H366.886ZM369.564 55.8902C369.243 55.8902 368.969 55.8028 368.743 55.6281C368.518 55.452 368.345 55.197 368.226 54.8632C368.106 54.5279 368.047 54.1231 368.047 53.6487C368.047 53.1771 368.106 52.7744 368.226 52.4406C368.346 52.1054 368.52 51.8497 368.746 51.6735C368.973 51.496 369.246 51.4072 369.564 51.4072C369.882 51.4072 370.154 51.496 370.38 51.6735C370.607 51.8497 370.78 52.1054 370.9 52.4406C371.02 52.7744 371.081 53.1771 371.081 53.6487C371.081 54.1231 371.021 54.5279 370.902 54.8632C370.783 55.197 370.61 55.452 370.384 55.6281C370.158 55.8028 369.885 55.8902 369.564 55.8902ZM369.564 55.4214C369.882 55.4214 370.129 55.268 370.305 54.9612C370.481 54.6544 370.569 54.2169 370.569 53.6487C370.569 53.2708 370.529 52.9491 370.448 52.6835C370.368 52.4179 370.253 52.2154 370.103 52.0762C369.954 51.937 369.774 51.8674 369.564 51.8674C369.248 51.8674 369.002 52.023 368.824 52.3341C368.647 52.6437 368.558 53.0819 368.558 53.6487C368.558 54.0265 368.598 54.3476 368.677 54.6118C368.757 54.876 368.871 55.077 369.02 55.2147C369.171 55.3525 369.352 55.4214 369.564 55.4214Z"
-        fill="black"
-      />
-      <path
-        id="15_2"
-        d="M379.967 46.4514V50.8151H379.439V47.0054H379.413L378.348 47.7128V47.1759L379.439 46.4514H379.967ZM382.586 50.8747C382.336 50.8747 382.11 50.825 381.91 50.7256C381.71 50.6261 381.549 50.4898 381.429 50.3165C381.308 50.1432 381.242 49.9457 381.23 49.7242H381.742C381.762 49.9216 381.851 50.0849 382.01 50.2142C382.171 50.3421 382.363 50.406 382.586 50.406C382.765 50.406 382.924 50.3641 383.063 50.2803C383.204 50.1965 383.314 50.0814 383.393 49.9351C383.474 49.7874 383.515 49.6205 383.515 49.4344C383.515 49.244 383.473 49.0743 383.389 48.9251C383.306 48.7746 383.193 48.656 383.048 48.5693C382.903 48.4827 382.738 48.4386 382.552 48.4372C382.418 48.4358 382.281 48.4564 382.14 48.499C382 48.5402 381.884 48.5935 381.793 48.6588L381.299 48.5992L381.563 46.4514H383.83V46.9202H382.006L381.853 48.2071H381.878C381.968 48.1361 382.08 48.0771 382.215 48.0303C382.35 47.9834 382.49 47.9599 382.637 47.9599C382.904 47.9599 383.142 48.0239 383.351 48.1517C383.561 48.2781 383.726 48.4514 383.845 48.6716C383.966 48.8918 384.026 49.1432 384.026 49.4259C384.026 49.7043 383.963 49.9528 383.838 50.1716C383.715 50.3889 383.544 50.5608 383.327 50.6872C383.11 50.8122 382.863 50.8747 382.586 50.8747Z"
-        fill="black"
-      />
-      <path
-        id="20_2"
-        d="M389.502 55.8305V55.447L390.942 53.8703C391.111 53.6856 391.25 53.5251 391.36 53.3887C391.469 53.251 391.55 53.1217 391.603 53.001C391.657 52.8788 391.684 52.751 391.684 52.6174C391.684 52.464 391.647 52.3312 391.573 52.219C391.5 52.1068 391.401 52.0201 391.275 51.9591C391.148 51.898 391.006 51.8674 390.848 51.8674C390.681 51.8674 390.534 51.9022 390.409 51.9718C390.286 52.04 390.19 52.1359 390.122 52.2595C390.055 52.3831 390.022 52.5279 390.022 52.6941H389.519C389.519 52.4385 389.578 52.214 389.696 52.0208C389.814 51.8277 389.974 51.6771 390.177 51.5691C390.382 51.4612 390.611 51.4072 390.865 51.4072C391.121 51.4072 391.348 51.4612 391.545 51.5691C391.743 51.6771 391.897 51.8227 392.01 52.0059C392.122 52.1892 392.178 52.393 392.178 52.6174C392.178 52.7779 392.149 52.9349 392.091 53.0883C392.034 53.2403 391.934 53.4101 391.792 53.5976C391.652 53.7836 391.456 54.0109 391.206 54.2794L390.226 55.3277V55.3618H392.255V55.8305H389.502ZM394.558 55.8902C394.237 55.8902 393.963 55.8028 393.738 55.6281C393.512 55.452 393.339 55.197 393.22 54.8632C393.101 54.5279 393.041 54.1231 393.041 53.6487C393.041 53.1771 393.101 52.7744 393.22 52.4406C393.341 52.1054 393.514 51.8497 393.74 51.6735C393.967 51.496 394.24 51.4072 394.558 51.4072C394.876 51.4072 395.148 51.496 395.374 51.6735C395.601 51.8497 395.775 52.1054 395.894 52.4406C396.015 52.7744 396.075 53.1771 396.075 53.6487C396.075 54.1231 396.015 54.5279 395.896 54.8632C395.777 55.197 395.604 55.452 395.378 55.6281C395.152 55.8028 394.879 55.8902 394.558 55.8902ZM394.558 55.4214C394.876 55.4214 395.123 55.268 395.299 54.9612C395.476 54.6544 395.564 54.2169 395.564 53.6487C395.564 53.2708 395.523 52.9491 395.442 52.6835C395.363 52.4179 395.248 52.2154 395.097 52.0762C394.948 51.937 394.768 51.8674 394.558 51.8674C394.243 51.8674 393.996 52.023 393.819 52.3341C393.641 52.6437 393.552 53.0819 393.552 53.6487C393.552 54.0265 393.592 54.3476 393.672 54.6118C393.751 54.876 393.865 55.077 394.015 55.2147C394.165 55.3525 394.346 55.4214 394.558 55.4214Z"
-        fill="black"
-      />
-      <path
-        id="25"
-        d="M395.539 68.8705V68.487L396.98 66.9103C397.149 66.7257 397.288 66.5651 397.397 66.4288C397.507 66.291 397.588 66.1617 397.64 66.041C397.694 65.9188 397.721 65.791 397.721 65.6575C397.721 65.5041 397.684 65.3713 397.61 65.259C397.538 65.1468 397.439 65.0602 397.312 64.9991C397.186 64.938 397.044 64.9075 396.886 64.9075C396.718 64.9075 396.572 64.9423 396.447 65.0119C396.324 65.0801 396.228 65.1759 396.159 65.2995C396.093 65.4231 396.059 65.568 396.059 65.7342H395.556C395.556 65.4785 395.615 65.2541 395.733 65.0609C395.851 64.8677 396.012 64.7171 396.215 64.6092C396.419 64.5012 396.649 64.4472 396.903 64.4472C397.159 64.4472 397.385 64.5012 397.583 64.6092C397.78 64.7171 397.935 64.8627 398.047 65.046C398.159 65.2292 398.216 65.433 398.216 65.6575C398.216 65.818 398.186 65.9749 398.128 66.1284C398.071 66.2803 397.972 66.4501 397.83 66.6376C397.689 66.8237 397.494 67.0509 397.244 67.3194L396.264 68.3677V68.4018H398.292V68.8705H395.539ZM400.536 68.9302C400.286 68.9302 400.061 68.8805 399.86 68.7811C399.66 68.6816 399.5 68.5453 399.379 68.372C399.258 68.1987 399.192 68.0012 399.181 67.7796H399.692C399.712 67.9771 399.801 68.1404 399.961 68.2697C400.121 68.3975 400.313 68.4615 400.536 68.4615C400.715 68.4615 400.874 68.4196 401.013 68.3357C401.154 68.2519 401.264 68.1369 401.343 67.9906C401.424 67.8428 401.465 67.6759 401.465 67.4899C401.465 67.2995 401.423 67.1298 401.339 66.9806C401.257 66.8301 401.143 66.7115 400.998 66.6248C400.853 66.5382 400.688 66.4941 400.502 66.4927C400.368 66.4913 400.231 66.5119 400.091 66.5545C399.95 66.5957 399.834 66.649 399.743 66.7143L399.249 66.6546L399.513 64.5069H401.78V64.9757H399.956L399.803 66.2626H399.828C399.918 66.1916 400.03 66.1326 400.165 66.0857C400.3 66.0389 400.441 66.0154 400.587 66.0154C400.854 66.0154 401.092 66.0794 401.301 66.2072C401.511 66.3336 401.676 66.5069 401.795 66.7271C401.916 66.9472 401.976 67.1987 401.976 67.4813C401.976 67.7597 401.914 68.0083 401.789 68.2271C401.665 68.4444 401.495 68.6163 401.277 68.7427C401.06 68.8677 400.813 68.9302 400.536 68.9302Z"
-        fill="black"
-      />
-      <path
-        id="30"
-        d="M390.976 81.9703C390.695 81.9703 390.444 81.922 390.224 81.8254C390.005 81.7288 389.831 81.5946 389.702 81.4227C389.574 81.2494 389.505 81.0484 389.493 80.8197H390.03C390.042 80.9604 390.09 81.0818 390.175 81.1841C390.26 81.2849 390.372 81.3631 390.51 81.4185C390.647 81.4739 390.8 81.5016 390.968 81.5016C391.155 81.5016 391.321 81.4689 391.466 81.4035C391.611 81.3382 391.725 81.2473 391.807 81.1308C391.89 81.0143 391.931 80.8794 391.931 80.726C391.931 80.5655 391.891 80.4241 391.811 80.302C391.732 80.1784 391.615 80.0818 391.462 80.0122C391.309 79.9426 391.121 79.9078 390.9 79.9078H390.55V79.4391H390.9C391.073 79.4391 391.225 79.4078 391.356 79.3453C391.488 79.2828 391.591 79.1947 391.664 79.0811C391.74 78.9675 391.777 78.8339 391.777 78.6805C391.777 78.5328 391.745 78.4043 391.679 78.2949C391.614 78.1855 391.522 78.1003 391.402 78.0392C391.284 77.9781 391.145 77.9476 390.985 77.9476C390.834 77.9476 390.692 77.9753 390.559 78.0307C390.427 78.0847 390.319 78.1635 390.235 78.2672C390.151 78.3695 390.106 78.493 390.098 78.6379H389.587C389.596 78.4092 389.664 78.2089 389.794 78.0371C389.923 77.8638 390.092 77.7288 390.301 77.6322C390.511 77.5356 390.742 77.4873 390.993 77.4873C391.263 77.4873 391.495 77.542 391.688 77.6514C391.881 77.7594 392.03 77.9021 392.133 78.0797C392.237 78.2572 392.289 78.449 392.289 78.655C392.289 78.9007 392.224 79.1102 392.095 79.2835C391.967 79.4568 391.793 79.5768 391.573 79.6436V79.6777C391.848 79.7231 392.064 79.8403 392.218 80.0293C392.373 80.2168 392.451 80.449 392.451 80.726C392.451 80.9632 392.386 81.1763 392.257 81.3652C392.129 81.5527 391.954 81.7004 391.733 81.8084C391.511 81.9163 391.259 81.9703 390.976 81.9703ZM394.745 81.9703C394.424 81.9703 394.151 81.8829 393.925 81.7082C393.699 81.5321 393.527 81.2771 393.407 80.9433C393.288 80.6081 393.228 80.2033 393.228 79.7288C393.228 79.2572 393.288 78.8545 393.407 78.5207C393.528 78.1855 393.701 77.9298 393.927 77.7537C394.155 77.5761 394.427 77.4873 394.745 77.4873C395.064 77.4873 395.336 77.5761 395.561 77.7537C395.789 77.9298 395.962 78.1855 396.081 78.5207C396.202 78.8545 396.262 79.2572 396.262 79.7288C396.262 80.2033 396.203 80.6081 396.083 80.9433C395.964 81.2771 395.792 81.5321 395.566 81.7082C395.34 81.8829 395.066 81.9703 394.745 81.9703ZM394.745 81.5016C395.064 81.5016 395.311 81.3481 395.487 81.0413C395.663 80.7345 395.751 80.297 395.751 79.7288C395.751 79.351 395.711 79.0293 395.63 78.7636C395.55 78.498 395.435 78.2956 395.284 78.1564C395.135 78.0172 394.956 77.9476 394.745 77.9476C394.43 77.9476 394.184 78.1031 394.006 78.4142C393.829 78.7239 393.74 79.1621 393.74 79.7288C393.74 80.1067 393.78 80.4277 393.859 80.6919C393.939 80.9561 394.053 81.1571 394.202 81.2949C394.353 81.4327 394.534 81.5016 394.745 81.5016Z"
-        fill="black"
-      />
-      <path
-        id="needle"
-        d="M381 72.9014C381.107 72.3497 381.196 71.798 381.269 71.2463C381.423 70.0723 381.5 68.8983 381.5 67.7244C381.5 63.5573 381.5 59.3902 381.5 55.2231C381.5 50.6336 381.367 46.0442 381.102 41.4548C381.07 40.9031 381.036 40.3514 381 39.7997C380.964 40.3514 380.93 40.9031 380.898 41.4548C380.633 46.0442 380.5 50.6336 380.5 55.2231C380.5 59.3902 380.5 63.5573 380.5 67.7244C380.5 68.8983 380.577 70.0723 380.731 71.2463C380.804 71.798 380.893 72.3497 381 72.9014Z"
-        fill="#E00000"
-      />
-      <path
-        id="pGaugeCenter"
-        d="M381 66.1298C381.418 66.1298 381.756 66.4677 381.756 66.8828C381.756 67.2979 381.418 67.6357 381 67.6357C380.581 67.6355 380.244 67.2978 380.244 66.8828C380.244 66.4678 380.581 66.13 381 66.1298Z"
-        fill="#D9D9D9"
-        stroke="black"
-        stroke-width="0.5"
-      />
-    </g>
-    <g id="ballValve">
-      <g id="body_7">
-        <rect
-          id="Rectangle 48_8"
-          x="60.5"
-          y="266.5"
-          width="29"
-          height="23"
-          fill="url(#paint80_linear_6_626)"
-          stroke="black"
-        />
-        <rect
-          id="Rectangle 49_7"
-          x="87.5"
-          y="263.5"
-          width="14"
-          height="29"
-          rx="0.5"
-          fill="url(#paint81_linear_6_626)"
-          stroke="black"
-        />
-        <rect
-          id="Rectangle 50_8"
-          x="48.5"
-          y="263.5"
-          width="14"
-          height="29"
-          rx="0.5"
-          fill="url(#paint82_linear_6_626)"
-          stroke="black"
-        />
-      </g>
-      <g id="steamValve">
-        <rect
-          id="Rectangle 51_8"
-          x="69.5"
-          y="266.5"
-          width="11"
-          height="17"
-          rx="0.5"
-          fill="#D9D9D9"
-          stroke="black"
-        />
-        <path
-          id="steamValveCenter"
-          d="M76.3565 280.348H73.6445L72.2891 278L73.6445 275.652H76.3565L77.7119 278L76.3565 280.348Z"
-          fill="#CECECE"
-          stroke="black"
-          stroke-width="0.5"
-        />
-        <path
-          id="Rectangle 52_5"
-          d="M72 189.5H78C81.0375 189.5 83.5 191.962 83.5 195V267C83.5 267.828 82.8284 268.5 82 268.5H68C67.1716 268.5 66.5 267.828 66.5 267V195C66.5 191.962 68.9624 189.5 72 189.5Z"
-          fill="#EE6300"
-          stroke="black"
-        />
-      </g>
-    </g>
-    <g id="ballValve_2">
-      <g id="body_8">
-        <rect
-          id="Rectangle 48_9"
-          x="509.5"
-          y="614.5"
-          width="29"
-          height="23"
-          transform="rotate(90 509.5 614.5)"
-          fill="url(#paint83_linear_6_626)"
-          stroke="black"
-        />
-        <rect
-          id="Rectangle 49_8"
-          x="512.5"
-          y="641.5"
-          width="14"
-          height="29"
-          rx="0.5"
-          transform="rotate(90 512.5 641.5)"
-          fill="url(#paint84_linear_6_626)"
-          stroke="black"
-        />
-        <rect
-          id="Rectangle 50_9"
-          x="512.5"
-          y="602.5"
-          width="14"
-          height="29"
-          rx="0.5"
-          transform="rotate(90 512.5 602.5)"
-          fill="url(#paint85_linear_6_626)"
-          stroke="black"
-        />
-      </g>
-      <g id="feedValve">
-        <path
-          id="valve"
-          fill-rule="evenodd"
-          clip-rule="evenodd"
-          d="M514.391 610.729C515.395 610.809 516.192 611.606 516.272 612.61L516.901 620.488C516.936 620.919 517.104 621.327 517.384 621.656L522.51 627.67C523.164 628.437 523.164 629.565 522.51 630.332L517.384 636.346C517.105 636.675 516.936 637.083 516.901 637.514L516.272 645.392C516.192 646.396 515.395 647.193 514.391 647.273L506.513 647.902C506.083 647.936 505.674 648.105 505.346 648.385L499.332 653.511L499.183 653.626C498.424 654.161 497.388 654.123 496.67 653.511L490.655 648.385C490.327 648.105 489.918 647.936 489.488 647.902L481.61 647.273L481.424 647.25C480.57 647.103 479.898 646.431 479.751 645.577L479.729 645.391L479.1 637.513C479.07 637.137 478.937 636.777 478.716 636.473L478.617 636.346L473.49 630.332C472.837 629.565 472.837 628.437 473.49 627.67L478.617 621.656C478.861 621.368 479.022 621.019 479.081 620.648L479.1 620.488L479.729 612.61C479.809 611.606 480.606 610.809 481.61 610.729L489.488 610.1C489.864 610.07 490.224 609.937 490.528 609.717L490.655 609.617L496.669 604.491C497.436 603.837 498.565 603.837 499.331 604.491L505.346 609.617C505.633 609.862 505.982 610.022 506.353 610.082L506.513 610.1L514.391 610.729ZM483.148 636.818C483.154 636.861 483.161 636.903 483.166 636.946L483.19 637.188L483.519 641.306L487.546 637.279C487.094 636.709 486.695 636.112 486.349 635.493L483.148 636.818ZM491.508 640.651C490.889 640.306 490.292 639.907 489.722 639.455L485.696 643.482L489.814 643.812L490.055 643.835C490.098 643.84 490.14 643.846 490.183 643.852L491.508 640.651ZM479.06 630.539L481.74 633.685C481.82 633.779 481.897 633.876 481.971 633.975L485.173 632.647C484.977 631.955 484.84 631.25 484.758 630.539L479.06 630.539ZM496.462 642.242C495.751 642.16 495.046 642.024 494.353 641.828L493.027 645.03C493.125 645.104 493.222 645.181 493.316 645.262L496.462 647.941L496.462 642.242ZM481.97 624.026C481.944 624.061 481.919 624.096 481.893 624.13L481.74 624.317L479.058 627.462L484.758 627.462C484.841 626.751 484.976 626.045 485.173 625.353L481.97 624.026ZM501.648 641.828C500.955 642.024 500.25 642.162 499.539 642.244L499.54 647.941L502.685 645.262C502.779 645.181 502.876 645.104 502.974 645.03L501.648 641.828ZM498.795 631.971C498.274 632.109 497.727 632.109 497.206 631.971L491.917 637.261C495.523 639.924 500.478 639.924 504.084 637.261L498.795 631.971ZM489.741 622.917C487.078 626.523 487.078 631.478 489.741 635.084L495.03 629.795C494.892 629.275 494.892 628.727 495.03 628.207L489.741 622.917ZM483.189 620.814C483.18 620.937 483.165 621.06 483.148 621.182L486.349 622.508C486.695 621.888 487.094 621.291 487.546 620.722L483.518 616.694L483.189 620.814ZM506.279 639.455C505.709 639.907 505.112 640.306 504.493 640.651L505.819 643.853C505.941 643.836 506.064 643.822 506.187 643.812L510.306 643.483L506.279 639.455ZM490.182 614.147C490.06 614.165 489.937 614.18 489.814 614.19L485.694 614.518L489.722 618.547C490.292 618.095 490.889 617.694 491.509 617.348L490.182 614.147ZM500.971 628.206C501.11 628.727 501.109 629.275 500.97 629.795L506.26 635.085C508.923 631.478 508.924 626.523 506.261 622.916L500.971 628.206ZM504.084 620.741C500.478 618.078 495.523 618.078 491.917 620.741L497.206 626.031C497.727 625.892 498.275 625.892 498.795 626.03L504.084 620.741ZM509.652 635.493C509.306 636.112 508.907 636.709 508.455 637.278L512.482 641.305L512.811 637.188C512.821 637.064 512.836 636.941 512.853 636.819L509.652 635.493ZM493.316 612.74C493.222 612.82 493.124 612.897 493.026 612.971L494.353 616.172C495.046 615.976 495.751 615.839 496.461 615.757L496.461 610.058L493.316 612.74ZM511.244 630.539C511.162 631.249 511.024 631.955 510.828 632.647L514.03 633.974C514.105 633.876 514.181 633.779 514.261 633.685L516.941 630.539L511.244 630.539ZM502.974 612.97C502.876 612.896 502.779 612.82 502.685 612.74L499.54 610.058L499.539 615.758C500.249 615.84 500.955 615.976 501.648 616.173L502.974 612.97ZM510.828 625.353C511.024 626.046 511.16 626.751 511.243 627.462L516.942 627.461L514.261 624.317C514.181 624.222 514.105 624.125 514.03 624.026L510.828 625.353ZM506.187 614.19C506.064 614.18 505.941 614.165 505.819 614.147L504.493 617.349C505.112 617.695 505.71 618.094 506.279 618.546L510.307 614.519L506.187 614.19ZM508.455 620.722C508.907 621.292 509.307 621.889 509.652 622.508L512.853 621.182C512.836 621.06 512.821 620.937 512.811 620.814L512.483 616.695L508.455 620.722Z"
-          fill="#FF3B3B"
-          stroke="black"
-          stroke-linejoin="round"
-        />
-        <path
-          id="feedValveCenter"
-          d="M496.518 629.855V628.146L498 627.289L499.483 628.146V629.855L498 630.712L496.518 629.855Z"
-          fill="#CECECE"
-          stroke="black"
-          stroke-width="0.5"
-        />
-      </g>
-    </g>
-    <g id="scale">
-      <path
-        id="Rectangle 27"
-        d="M548.687 647.5H697.313C697.52 647.5 697.706 647.627 697.78 647.82L712.271 685.5H533.729L548.22 647.82C548.294 647.627 548.48 647.5 548.687 647.5Z"
-        fill="#CECECE"
-        stroke="black"
-      />
-      <rect
-        id="Rectangle 28"
-        x="533.5"
-        y="693.5"
-        width="179"
-        height="53"
-        rx="3.5"
-        fill="#888888"
-        stroke="black"
-      />
-      <rect
-        id="scaleScreen"
-        x="538.5"
-        y="699.5"
-        width="125"
-        height="41"
-        rx="1.5"
-        fill="#141414"
-        stroke="black"
-      />
-      <g id="tareBtn">
-        <rect
-          id="upBtnbase"
-          x="670.5"
-          y="702.5"
-          width="35"
-          height="35"
-          rx="5.5"
-          fill="#D51F1F"
-          stroke="black"
-        />
-        <path
-          id="0_3"
-          d="M688 733.44C685.52 733.44 683.613 732.333 682.28 730.12C680.947 727.907 680.28 724.76 680.28 720.68C680.28 716.733 680.947 713.627 682.28 711.36C683.613 709.067 685.52 707.92 688 707.92C690.48 707.92 692.387 709.067 693.72 711.36C695.053 713.627 695.72 716.733 695.72 720.68C695.72 724.76 695.053 727.907 693.72 730.12C692.387 732.333 690.48 733.44 688 733.44ZM683.16 720.68C683.16 722.387 683.28 723.907 683.52 725.24L692 714.32C691.6 713.067 691.053 712.093 690.36 711.4C689.693 710.707 688.907 710.36 688 710.36C686.48 710.36 685.293 711.293 684.44 713.16C683.587 715.027 683.16 717.533 683.16 720.68ZM688 731C689.52 731 690.707 730.107 691.56 728.32C692.413 726.533 692.84 723.987 692.84 720.68C692.84 719.213 692.747 717.88 692.56 716.68L684.08 727.44C684.907 729.813 686.213 731 688 731Z"
-          fill="#320101"
-        />
-      </g>
-      <rect
-        id="Rectangle 29"
-        x="533.5"
-        y="684.5"
-        width="179"
-        height="10"
-        rx="1.5"
-        fill="#CECECE"
-        stroke="black"
-      />
-    </g>
-    <g id="beaker2">
-      <rect
-        id="beaker2Fill"
-        x="583"
-        y="563"
-        width="79"
-        height="100"
-        fill="#597AFF"
-        fill-opacity="0.8"
-      />
-      <g id="beakerBody">
-        <g id="Rectangle 3">
-          <mask id="path-287-inside-6_6_626" fill="white">
-            <path
-              d="M579 547H666V661C666 664.314 663.314 667 660 667H585C581.686 667 579 664.314 579 661V547Z"
-            />
-          </mask>
-          <path
-            d="M579 547H666H579ZM670 661C670 666.523 665.523 671 660 671H585C579.477 671 575 666.523 575 661H583C583 662.105 583.895 663 585 663H660C661.105 663 662 662.105 662 661H670ZM585 671C579.477 671 575 666.523 575 661V547H583V661C583 662.105 583.895 663 585 663V671ZM670 547V661C670 666.523 665.523 671 660 671V663C661.105 663 662 662.105 662 661V547H670Z"
-            fill="#D5D5D5"
-            mask="url(#path-287-inside-6_6_626)"
-          />
-        </g>
-        <path
-          id="Rectangle 1"
-          d="M665.5 547.5V661C665.5 664.038 663.038 666.5 660 666.5H585C581.962 666.5 579.5 664.038 579.5 661V547.5H665.5Z"
-          fill="url(#paint86_linear_6_626)"
-          fill-opacity="0.2"
-          stroke="black"
-        />
-        <rect
-          id="Rectangle 2"
-          x="578.5"
-          y="545.5"
-          width="89"
-          height="3"
-          rx="1.5"
-          fill="#D5D5D5"
-          stroke="black"
-        />
-      </g>
-      <g id="ticks">
-        <g id="Repeat group 1_5">
-          <line
-            id="Line 22"
-            x1="610"
-            y1="565.25"
-            x2="600"
-            y2="565.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_2"
-            x1="610"
-            y1="567.25"
-            x2="600"
-            y2="567.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_3"
-            x1="610"
-            y1="569.25"
-            x2="600"
-            y2="569.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_4"
-            x1="610"
-            y1="571.25"
-            x2="600"
-            y2="571.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_5"
-            x1="610"
-            y1="575.25"
-            x2="600"
-            y2="575.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_6"
-            x1="610"
-            y1="577.25"
-            x2="600"
-            y2="577.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_7"
-            x1="610"
-            y1="579.25"
-            x2="600"
-            y2="579.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_8"
-            x1="610"
-            y1="581.25"
-            x2="600"
-            y2="581.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_9"
-            x1="610"
-            y1="585.25"
-            x2="600"
-            y2="585.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_10"
-            x1="610"
-            y1="587.25"
-            x2="600"
-            y2="587.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_11"
-            x1="610"
-            y1="589.25"
-            x2="600"
-            y2="589.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_12"
-            x1="610"
-            y1="591.25"
-            x2="600"
-            y2="591.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_13"
-            x1="610"
-            y1="595.25"
-            x2="600"
-            y2="595.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_14"
-            x1="610"
-            y1="597.25"
-            x2="600"
-            y2="597.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_15"
-            x1="610"
-            y1="599.25"
-            x2="600"
-            y2="599.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_16"
-            x1="610"
-            y1="601.25"
-            x2="600"
-            y2="601.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_17"
-            x1="610"
-            y1="605.25"
-            x2="600"
-            y2="605.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_18"
-            x1="610"
-            y1="607.25"
-            x2="600"
-            y2="607.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_19"
-            x1="610"
-            y1="609.25"
-            x2="600"
-            y2="609.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_20"
-            x1="610"
-            y1="611.25"
-            x2="600"
-            y2="611.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_21"
-            x1="610"
-            y1="615.25"
-            x2="600"
-            y2="615.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_22"
-            x1="610"
-            y1="617.25"
-            x2="600"
-            y2="617.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_23"
-            x1="610"
-            y1="619.25"
-            x2="600"
-            y2="619.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_24"
-            x1="610"
-            y1="621.25"
-            x2="600"
-            y2="621.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_25"
-            x1="610"
-            y1="625.25"
-            x2="600"
-            y2="625.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_26"
-            x1="610"
-            y1="627.25"
-            x2="600"
-            y2="627.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_27"
-            x1="610"
-            y1="629.25"
-            x2="600"
-            y2="629.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_28"
-            x1="610"
-            y1="631.25"
-            x2="600"
-            y2="631.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_29"
-            x1="610"
-            y1="635.25"
-            x2="600"
-            y2="635.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_30"
-            x1="610"
-            y1="637.25"
-            x2="600"
-            y2="637.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_31"
-            x1="610"
-            y1="639.25"
-            x2="600"
-            y2="639.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_32"
-            x1="610"
-            y1="641.25"
-            x2="600"
-            y2="641.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-        </g>
-        <g id="Group 5">
-          <line
-            id="Line 22_33"
-            x1="620"
-            y1="563.25"
-            x2="600"
-            y2="563.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_34"
-            x1="615"
-            y1="573.25"
-            x2="600"
-            y2="573.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_35"
-            x1="620"
-            y1="583.25"
-            x2="600"
-            y2="583.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_36"
-            x1="615"
-            y1="593.25"
-            x2="600"
-            y2="593.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_37"
-            x1="620"
-            y1="603.25"
-            x2="600"
-            y2="603.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_38"
-            x1="615"
-            y1="613.25"
-            x2="600"
-            y2="613.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_39"
-            x1="620"
-            y1="623.25"
-            x2="600"
-            y2="623.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_40"
-            x1="615"
-            y1="633.25"
-            x2="600"
-            y2="633.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 23"
-            x1="620"
-            y1="643.25"
-            x2="600"
-            y2="643.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-        </g>
-        <path
-          id="1000 mL"
-          d="M622.986 560.636V565H622.457V561.19H622.432L621.366 561.898V561.361L622.457 560.636H622.986ZM625.664 565.06C625.343 565.06 625.07 564.972 624.844 564.798C624.618 564.621 624.445 564.366 624.326 564.033C624.207 563.697 624.147 563.293 624.147 562.818C624.147 562.347 624.207 561.944 624.326 561.61C624.447 561.275 624.62 561.019 624.846 560.843C625.073 560.665 625.346 560.577 625.664 560.577C625.982 560.577 626.254 560.665 626.48 560.843C626.707 561.019 626.881 561.275 627 561.61C627.121 561.944 627.181 562.347 627.181 562.818C627.181 563.293 627.121 563.697 627.002 564.033C626.883 564.366 626.71 564.621 626.484 564.798C626.259 564.972 625.985 565.06 625.664 565.06ZM625.664 564.591C625.982 564.591 626.229 564.437 626.406 564.131C626.582 563.824 626.67 563.386 626.67 562.818C626.67 562.44 626.629 562.119 626.548 561.853C626.469 561.587 626.354 561.385 626.203 561.246C626.054 561.107 625.874 561.037 625.664 561.037C625.349 561.037 625.102 561.192 624.925 561.504C624.747 561.813 624.658 562.251 624.658 562.818C624.658 563.196 624.698 563.517 624.778 563.781C624.857 564.045 624.972 564.246 625.121 564.384C625.271 564.522 625.452 564.591 625.664 564.591ZM629.414 565.06C629.093 565.06 628.82 564.972 628.594 564.798C628.368 564.621 628.195 564.366 628.076 564.033C627.957 563.697 627.897 563.293 627.897 562.818C627.897 562.347 627.957 561.944 628.076 561.61C628.197 561.275 628.37 561.019 628.596 560.843C628.823 560.665 629.096 560.577 629.414 560.577C629.732 560.577 630.004 560.665 630.23 560.843C630.457 561.019 630.631 561.275 630.75 561.61C630.871 561.944 630.931 562.347 630.931 562.818C630.931 563.293 630.871 563.697 630.752 564.033C630.633 564.366 630.46 564.621 630.234 564.798C630.009 564.972 629.735 565.06 629.414 565.06ZM629.414 564.591C629.732 564.591 629.979 564.437 630.156 564.131C630.332 563.824 630.42 563.386 630.42 562.818C630.42 562.44 630.379 562.119 630.298 561.853C630.219 561.587 630.104 561.385 629.953 561.246C629.804 561.107 629.624 561.037 629.414 561.037C629.099 561.037 628.852 561.192 628.675 561.504C628.497 561.813 628.408 562.251 628.408 562.818C628.408 563.196 628.448 563.517 628.528 563.781C628.607 564.045 628.722 564.246 628.871 564.384C629.021 564.522 629.202 564.591 629.414 564.591ZM633.164 565.06C632.843 565.06 632.57 564.972 632.344 564.798C632.118 564.621 631.945 564.366 631.826 564.033C631.707 563.697 631.647 563.293 631.647 562.818C631.647 562.347 631.707 561.944 631.826 561.61C631.947 561.275 632.12 561.019 632.346 560.843C632.573 560.665 632.846 560.577 633.164 560.577C633.482 560.577 633.754 560.665 633.98 560.843C634.207 561.019 634.381 561.275 634.5 561.61C634.621 561.944 634.681 562.347 634.681 562.818C634.681 563.293 634.621 563.697 634.502 564.033C634.383 564.366 634.21 564.621 633.984 564.798C633.759 564.972 633.485 565.06 633.164 565.06ZM633.164 564.591C633.482 564.591 633.729 564.437 633.906 564.131C634.082 563.824 634.17 563.386 634.17 562.818C634.17 562.44 634.129 562.119 634.048 561.853C633.969 561.587 633.854 561.385 633.703 561.246C633.554 561.107 633.374 561.037 633.164 561.037C632.849 561.037 632.602 561.192 632.425 561.504C632.247 561.813 632.158 562.251 632.158 562.818C632.158 563.196 632.198 563.517 632.278 563.781C632.357 564.045 632.472 564.246 632.621 564.384C632.771 564.522 632.952 564.591 633.164 564.591ZM637.187 565V561.727H637.673V562.239H637.715C637.783 562.064 637.893 561.928 638.045 561.832C638.197 561.734 638.38 561.685 638.593 561.685C638.809 561.685 638.989 561.734 639.132 561.832C639.277 561.928 639.39 562.064 639.471 562.239H639.505C639.589 562.07 639.714 561.935 639.882 561.836C640.05 561.735 640.251 561.685 640.485 561.685C640.778 561.685 641.017 561.776 641.203 561.96C641.389 562.141 641.482 562.425 641.482 562.81V565H640.979V562.81C640.979 562.568 640.913 562.396 640.781 562.292C640.649 562.188 640.494 562.136 640.315 562.136C640.085 562.136 639.906 562.206 639.78 562.345C639.653 562.483 639.59 562.658 639.59 562.869V565H639.079V562.759C639.079 562.572 639.018 562.423 638.898 562.309C638.777 562.194 638.621 562.136 638.431 562.136C638.3 562.136 638.178 562.171 638.065 562.241C637.952 562.31 637.862 562.407 637.792 562.531C637.724 562.653 637.69 562.794 637.69 562.955V565H637.187ZM642.47 565V560.636H642.998V564.531H645.027V565H642.47Z"
-          fill="black"
-        />
-        <path
-          id="800"
-          d="M624.849 585.06C624.557 585.06 624.298 585.008 624.074 584.904C623.851 584.799 623.677 584.655 623.552 584.472C623.427 584.287 623.365 584.077 623.366 583.841C623.365 583.656 623.401 583.486 623.475 583.33C623.549 583.172 623.65 583.04 623.778 582.935C623.907 582.829 624.051 582.761 624.21 582.733V582.707C624.001 582.653 623.835 582.536 623.712 582.356C623.588 582.174 623.527 581.967 623.528 581.736C623.527 581.514 623.583 581.316 623.697 581.141C623.81 580.967 623.967 580.829 624.165 580.728C624.366 580.627 624.594 580.577 624.849 580.577C625.102 580.577 625.328 580.627 625.527 580.728C625.726 580.829 625.882 580.967 625.996 581.141C626.111 581.316 626.169 581.514 626.17 581.736C626.169 581.967 626.106 582.174 625.981 582.356C625.857 582.536 625.693 582.653 625.489 582.707V582.733C625.646 582.761 625.788 582.829 625.915 582.935C626.041 583.04 626.142 583.172 626.217 583.33C626.293 583.486 626.331 583.656 626.332 583.841C626.331 584.077 626.267 584.287 626.141 584.472C626.016 584.655 625.842 584.799 625.619 584.904C625.397 585.008 625.141 585.06 624.849 585.06ZM624.849 584.591C625.047 584.591 625.217 584.559 625.361 584.495C625.504 584.431 625.615 584.341 625.693 584.224C625.771 584.108 625.811 583.972 625.812 583.815C625.811 583.651 625.768 583.505 625.685 583.379C625.601 583.252 625.487 583.153 625.342 583.08C625.198 583.008 625.034 582.972 624.849 582.972C624.663 582.972 624.497 583.008 624.351 583.08C624.206 583.153 624.092 583.252 624.008 583.379C623.925 583.505 623.885 583.651 623.886 583.815C623.885 583.972 623.923 584.108 623.999 584.224C624.077 584.341 624.189 584.431 624.334 584.495C624.479 584.559 624.651 584.591 624.849 584.591ZM624.849 582.52C625.006 582.52 625.144 582.489 625.265 582.426C625.387 582.364 625.483 582.276 625.553 582.164C625.622 582.052 625.658 581.92 625.659 581.77C625.658 581.622 625.623 581.494 625.555 581.384C625.487 581.273 625.392 581.188 625.271 581.129C625.151 581.067 625.01 581.037 624.849 581.037C624.686 581.037 624.543 581.067 624.421 581.129C624.299 581.188 624.205 581.273 624.138 581.384C624.071 581.494 624.038 581.622 624.04 581.77C624.038 581.92 624.072 582.052 624.14 582.164C624.21 582.276 624.305 582.364 624.428 582.426C624.55 582.489 624.69 582.52 624.849 582.52ZM628.572 585.06C628.251 585.06 627.978 584.972 627.752 584.798C627.526 584.621 627.354 584.366 627.234 584.033C627.115 583.697 627.055 583.293 627.055 582.818C627.055 582.347 627.115 581.944 627.234 581.61C627.355 581.275 627.528 581.019 627.754 580.843C627.981 580.665 628.254 580.577 628.572 580.577C628.89 580.577 629.162 580.665 629.388 580.843C629.616 581.019 629.789 581.275 629.908 581.61C630.029 581.944 630.089 582.347 630.089 582.818C630.089 583.293 630.03 583.697 629.91 584.033C629.791 584.366 629.618 584.621 629.393 584.798C629.167 584.972 628.893 585.06 628.572 585.06ZM628.572 584.591C628.89 584.591 629.138 584.437 629.314 584.131C629.49 583.824 629.578 583.386 629.578 582.818C629.578 582.44 629.537 582.119 629.456 581.853C629.377 581.587 629.262 581.385 629.111 581.246C628.962 581.107 628.782 581.037 628.572 581.037C628.257 581.037 628.01 581.192 627.833 581.504C627.655 581.813 627.567 582.251 627.567 582.818C627.567 583.196 627.606 583.517 627.686 583.781C627.765 584.045 627.88 584.246 628.029 584.384C628.18 584.522 628.361 584.591 628.572 584.591ZM632.322 585.06C632.001 585.06 631.728 584.972 631.502 584.798C631.276 584.621 631.104 584.366 630.984 584.033C630.865 583.697 630.805 583.293 630.805 582.818C630.805 582.347 630.865 581.944 630.984 581.61C631.105 581.275 631.278 581.019 631.504 580.843C631.731 580.665 632.004 580.577 632.322 580.577C632.64 580.577 632.912 580.665 633.138 580.843C633.366 581.019 633.539 581.275 633.658 581.61C633.779 581.944 633.839 582.347 633.839 582.818C633.839 583.293 633.78 583.697 633.66 584.033C633.541 584.366 633.368 584.621 633.143 584.798C632.917 584.972 632.643 585.06 632.322 585.06ZM632.322 584.591C632.64 584.591 632.888 584.437 633.064 584.131C633.24 583.824 633.328 583.386 633.328 582.818C633.328 582.44 633.287 582.119 633.206 581.853C633.127 581.587 633.012 581.385 632.861 581.246C632.712 581.107 632.532 581.037 632.322 581.037C632.007 581.037 631.76 581.192 631.583 581.504C631.405 581.813 631.317 582.251 631.317 582.818C631.317 583.196 631.356 583.517 631.436 583.781C631.515 584.045 631.63 584.246 631.779 584.384C631.93 584.522 632.111 584.591 632.322 584.591Z"
-          fill="black"
-        />
-        <path
-          id="600"
-          d="M624.901 605.06C624.722 605.057 624.543 605.023 624.364 604.957C624.185 604.892 624.021 604.782 623.874 604.627C623.726 604.471 623.607 604.26 623.518 603.994C623.428 603.727 623.384 603.392 623.384 602.989C623.384 602.602 623.42 602.26 623.492 601.962C623.565 601.662 623.67 601.41 623.808 601.205C623.945 600.999 624.112 600.843 624.306 600.737C624.502 600.63 624.723 600.577 624.969 600.577C625.213 600.577 625.43 600.626 625.621 600.724C625.813 600.82 625.969 600.955 626.089 601.129C626.21 601.302 626.288 601.501 626.324 601.727H625.804C625.756 601.531 625.662 601.369 625.523 601.239C625.384 601.11 625.199 601.045 624.969 601.045C624.631 601.045 624.364 601.192 624.17 601.487C623.977 601.781 623.879 602.193 623.878 602.724H623.912C623.991 602.604 624.086 602.501 624.195 602.415C624.306 602.329 624.428 602.262 624.562 602.215C624.695 602.168 624.837 602.145 624.986 602.145C625.236 602.145 625.464 602.207 625.672 602.332C625.879 602.456 626.045 602.627 626.17 602.846C626.295 603.063 626.358 603.312 626.358 603.594C626.358 603.864 626.298 604.111 626.177 604.335C626.056 604.558 625.886 604.736 625.668 604.868C625.45 604.999 625.195 605.062 624.901 605.06ZM624.901 604.591C625.08 604.591 625.24 604.546 625.382 604.457C625.526 604.367 625.638 604.247 625.721 604.097C625.805 603.946 625.847 603.778 625.847 603.594C625.847 603.413 625.806 603.249 625.725 603.102C625.646 602.952 625.536 602.834 625.395 602.746C625.256 602.658 625.097 602.614 624.918 602.614C624.783 602.614 624.657 602.641 624.54 602.695C624.424 602.747 624.322 602.82 624.234 602.912C624.147 603.004 624.079 603.11 624.029 603.229C623.979 603.347 623.955 603.472 623.955 603.602C623.955 603.776 623.995 603.937 624.076 604.088C624.158 604.239 624.271 604.36 624.413 604.452C624.556 604.545 624.719 604.591 624.901 604.591ZM628.619 605.06C628.298 605.06 628.025 604.972 627.799 604.798C627.573 604.621 627.4 604.366 627.281 604.033C627.162 603.697 627.102 603.293 627.102 602.818C627.102 602.347 627.162 601.944 627.281 601.61C627.402 601.275 627.575 601.019 627.801 600.843C628.028 600.665 628.301 600.577 628.619 600.577C628.937 600.577 629.209 600.665 629.435 600.843C629.662 601.019 629.836 601.275 629.955 601.61C630.076 601.944 630.136 602.347 630.136 602.818C630.136 603.293 630.077 603.697 629.957 604.033C629.838 604.366 629.665 604.621 629.439 604.798C629.214 604.972 628.94 605.06 628.619 605.06ZM628.619 604.591C628.937 604.591 629.184 604.437 629.361 604.131C629.537 603.824 629.625 603.386 629.625 602.818C629.625 602.44 629.584 602.119 629.503 601.853C629.424 601.587 629.309 601.385 629.158 601.246C629.009 601.107 628.829 601.037 628.619 601.037C628.304 601.037 628.057 601.192 627.88 601.504C627.702 601.813 627.613 602.251 627.613 602.818C627.613 603.196 627.653 603.517 627.733 603.781C627.812 604.045 627.927 604.246 628.076 604.384C628.226 604.522 628.407 604.591 628.619 604.591ZM632.369 605.06C632.048 605.06 631.775 604.972 631.549 604.798C631.323 604.621 631.15 604.366 631.031 604.033C630.912 603.697 630.852 603.293 630.852 602.818C630.852 602.347 630.912 601.944 631.031 601.61C631.152 601.275 631.325 601.019 631.551 600.843C631.778 600.665 632.051 600.577 632.369 600.577C632.687 600.577 632.959 600.665 633.185 600.843C633.412 601.019 633.586 601.275 633.705 601.61C633.826 601.944 633.886 602.347 633.886 602.818C633.886 603.293 633.827 603.697 633.707 604.033C633.588 604.366 633.415 604.621 633.189 604.798C632.964 604.972 632.69 605.06 632.369 605.06ZM632.369 604.591C632.687 604.591 632.934 604.437 633.111 604.131C633.287 603.824 633.375 603.386 633.375 602.818C633.375 602.44 633.334 602.119 633.253 601.853C633.174 601.587 633.059 601.385 632.908 601.246C632.759 601.107 632.579 601.037 632.369 601.037C632.054 601.037 631.807 601.192 631.63 601.504C631.452 601.813 631.363 602.251 631.363 602.818C631.363 603.196 631.403 603.517 631.483 603.781C631.562 604.045 631.677 604.246 631.826 604.384C631.976 604.522 632.157 604.591 632.369 604.591Z"
-          fill="black"
-        />
-        <path
-          id="400"
-          d="M623.349 624.105V623.67L625.267 620.636H625.582V621.31H625.369L623.92 623.602V623.636H626.503V624.105H623.349ZM625.403 625V623.973V623.771V620.636H625.906V625H625.403ZM628.725 625.06C628.404 625.06 628.13 624.972 627.904 624.798C627.678 624.621 627.506 624.366 627.387 624.033C627.267 623.697 627.208 623.293 627.208 622.818C627.208 622.347 627.267 621.944 627.387 621.61C627.507 621.275 627.681 621.019 627.906 620.843C628.134 620.665 628.406 620.577 628.725 620.577C629.043 620.577 629.315 620.665 629.541 620.843C629.768 621.019 629.941 621.275 630.061 621.61C630.181 621.944 630.242 622.347 630.242 622.818C630.242 623.293 630.182 623.697 630.063 624.033C629.943 624.366 629.771 624.621 629.545 624.798C629.319 624.972 629.046 625.06 628.725 625.06ZM628.725 624.591C629.043 624.591 629.29 624.437 629.466 624.131C629.642 623.824 629.73 623.386 629.73 622.818C629.73 622.44 629.69 622.119 629.609 621.853C629.529 621.587 629.414 621.385 629.264 621.246C629.115 621.107 628.935 621.037 628.725 621.037C628.409 621.037 628.163 621.192 627.985 621.504C627.808 621.813 627.719 622.251 627.719 622.818C627.719 623.196 627.759 623.517 627.838 623.781C627.918 624.045 628.032 624.246 628.181 624.384C628.332 624.522 628.513 624.591 628.725 624.591ZM632.475 625.06C632.154 625.06 631.88 624.972 631.654 624.798C631.428 624.621 631.256 624.366 631.137 624.033C631.017 623.697 630.958 623.293 630.958 622.818C630.958 622.347 631.017 621.944 631.137 621.61C631.257 621.275 631.431 621.019 631.656 620.843C631.884 620.665 632.156 620.577 632.475 620.577C632.793 620.577 633.065 620.665 633.291 620.843C633.518 621.019 633.691 621.275 633.811 621.61C633.931 621.944 633.992 622.347 633.992 622.818C633.992 623.293 633.932 623.697 633.813 624.033C633.693 624.366 633.521 624.621 633.295 624.798C633.069 624.972 632.796 625.06 632.475 625.06ZM632.475 624.591C632.793 624.591 633.04 624.437 633.216 624.131C633.392 623.824 633.48 623.386 633.48 622.818C633.48 622.44 633.44 622.119 633.359 621.853C633.279 621.587 633.164 621.385 633.014 621.246C632.865 621.107 632.685 621.037 632.475 621.037C632.159 621.037 631.913 621.192 631.735 621.504C631.558 621.813 631.469 622.251 631.469 622.818C631.469 623.196 631.509 623.517 631.588 623.781C631.668 624.045 631.782 624.246 631.931 624.384C632.082 624.522 632.263 624.591 632.475 624.591Z"
-          fill="black"
-        />
-        <path
-          id="200"
-          d="M623.452 645V644.616L624.892 643.04C625.061 642.855 625.2 642.695 625.31 642.558C625.419 642.42 625.5 642.291 625.553 642.17C625.607 642.048 625.634 641.92 625.634 641.787C625.634 641.634 625.597 641.501 625.523 641.388C625.45 641.276 625.351 641.19 625.224 641.129C625.098 641.067 624.956 641.037 624.798 641.037C624.631 641.037 624.484 641.072 624.359 641.141C624.236 641.21 624.14 641.305 624.072 641.429C624.005 641.553 623.972 641.697 623.972 641.864H623.469C623.469 641.608 623.528 641.384 623.646 641.19C623.763 640.997 623.924 640.847 624.127 640.739C624.332 640.631 624.561 640.577 624.815 640.577C625.071 640.577 625.298 640.631 625.495 640.739C625.692 640.847 625.847 640.992 625.96 641.175C626.072 641.359 626.128 641.562 626.128 641.787C626.128 641.947 626.099 642.104 626.04 642.258C625.984 642.41 625.884 642.58 625.742 642.767C625.602 642.953 625.406 643.18 625.156 643.449L624.176 644.497V644.531H626.205V645H623.452ZM628.508 645.06C628.187 645.06 627.913 644.972 627.688 644.798C627.462 644.621 627.289 644.366 627.17 644.033C627.05 643.697 626.991 643.293 626.991 642.818C626.991 642.347 627.05 641.944 627.17 641.61C627.29 641.275 627.464 641.019 627.69 640.843C627.917 640.665 628.19 640.577 628.508 640.577C628.826 640.577 629.098 640.665 629.324 640.843C629.551 641.019 629.724 641.275 629.844 641.61C629.964 641.944 630.025 642.347 630.025 642.818C630.025 643.293 629.965 643.697 629.846 644.033C629.727 644.366 629.554 644.621 629.328 644.798C629.102 644.972 628.829 645.06 628.508 645.06ZM628.508 644.591C628.826 644.591 629.073 644.437 629.249 644.131C629.425 643.824 629.513 643.386 629.513 642.818C629.513 642.44 629.473 642.119 629.392 641.853C629.313 641.587 629.197 641.385 629.047 641.246C628.898 641.107 628.718 641.037 628.508 641.037C628.192 641.037 627.946 641.192 627.768 641.504C627.591 641.813 627.502 642.251 627.502 642.818C627.502 643.196 627.542 643.517 627.621 643.781C627.701 644.045 627.815 644.246 627.964 644.384C628.115 644.522 628.296 644.591 628.508 644.591ZM632.258 645.06C631.937 645.06 631.663 644.972 631.438 644.798C631.212 644.621 631.039 644.366 630.92 644.033C630.8 643.697 630.741 643.293 630.741 642.818C630.741 642.347 630.8 641.944 630.92 641.61C631.04 641.275 631.214 641.019 631.44 640.843C631.667 640.665 631.94 640.577 632.258 640.577C632.576 640.577 632.848 640.665 633.074 640.843C633.301 641.019 633.474 641.275 633.594 641.61C633.714 641.944 633.775 642.347 633.775 642.818C633.775 643.293 633.715 643.697 633.596 644.033C633.477 644.366 633.304 644.621 633.078 644.798C632.852 644.972 632.579 645.06 632.258 645.06ZM632.258 644.591C632.576 644.591 632.823 644.437 632.999 644.131C633.175 643.824 633.263 643.386 633.263 642.818C633.263 642.44 633.223 642.119 633.142 641.853C633.063 641.587 632.947 641.385 632.797 641.246C632.648 641.107 632.468 641.037 632.258 641.037C631.942 641.037 631.696 641.192 631.518 641.504C631.341 641.813 631.252 642.251 631.252 642.818C631.252 643.196 631.292 643.517 631.371 643.781C631.451 644.045 631.565 644.246 631.714 644.384C631.865 644.522 632.046 644.591 632.258 644.591Z"
-          fill="black"
-        />
-        <line
-          id="Line 24"
-          x1="600.25"
-          y1="563"
-          x2="600.25"
-          y2="643"
-          stroke="black"
-          stroke-width="0.5"
-        />
-      </g>
-    </g>
-    <g id="beaker1">
-      <rect
-        id="beaker1Fill"
-        x="282"
-        y="630"
-        width="79"
-        height="100"
-        fill="#597AFF"
-        fill-opacity="0.8"
-      />
-      <g id="beakerBody_2">
-        <g id="Rectangle 3_2">
-          <mask id="path-339-inside-7_6_626" fill="white">
-            <path
-              d="M278 614H365V728C365 731.314 362.314 734 359 734H284C280.686 734 278 731.314 278 728V614Z"
-            />
-          </mask>
-          <path
-            d="M278 614H365H278ZM369 728C369 733.523 364.523 738 359 738H284C278.477 738 274 733.523 274 728H282C282 729.105 282.895 730 284 730H359C360.105 730 361 729.105 361 728H369ZM284 738C278.477 738 274 733.523 274 728V614H282V728C282 729.105 282.895 730 284 730V738ZM369 614V728C369 733.523 364.523 738 359 738V730C360.105 730 361 729.105 361 728V614H369Z"
-            fill="#D5D5D5"
-            mask="url(#path-339-inside-7_6_626)"
-          />
-        </g>
-        <path
-          id="Rectangle 1_2"
-          d="M364.5 614.5V728C364.5 731.038 362.038 733.5 359 733.5H284C280.962 733.5 278.5 731.038 278.5 728V614.5H364.5Z"
-          fill="url(#paint87_linear_6_626)"
-          fill-opacity="0.2"
-          stroke="black"
-        />
-        <rect
-          id="Rectangle 2_2"
-          x="277.5"
-          y="612.5"
-          width="89"
-          height="3"
-          rx="1.5"
-          fill="#D5D5D5"
-          stroke="black"
-        />
-      </g>
-      <g id="ticks_2">
-        <g id="Repeat group 1_6">
-          <line
-            id="Line 22_41"
-            x1="309"
-            y1="632.25"
-            x2="299"
-            y2="632.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_42"
-            x1="309"
-            y1="634.25"
-            x2="299"
-            y2="634.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_43"
-            x1="309"
-            y1="636.25"
-            x2="299"
-            y2="636.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_44"
-            x1="309"
-            y1="638.25"
-            x2="299"
-            y2="638.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_45"
-            x1="309"
-            y1="642.25"
-            x2="299"
-            y2="642.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_46"
-            x1="309"
-            y1="644.25"
-            x2="299"
-            y2="644.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_47"
-            x1="309"
-            y1="646.25"
-            x2="299"
-            y2="646.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_48"
-            x1="309"
-            y1="648.25"
-            x2="299"
-            y2="648.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_49"
-            x1="309"
-            y1="652.25"
-            x2="299"
-            y2="652.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_50"
-            x1="309"
-            y1="654.25"
-            x2="299"
-            y2="654.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_51"
-            x1="309"
-            y1="656.25"
-            x2="299"
-            y2="656.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_52"
-            x1="309"
-            y1="658.25"
-            x2="299"
-            y2="658.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_53"
-            x1="309"
-            y1="662.25"
-            x2="299"
-            y2="662.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_54"
-            x1="309"
-            y1="664.25"
-            x2="299"
-            y2="664.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_55"
-            x1="309"
-            y1="666.25"
-            x2="299"
-            y2="666.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_56"
-            x1="309"
-            y1="668.25"
-            x2="299"
-            y2="668.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_57"
-            x1="309"
-            y1="672.25"
-            x2="299"
-            y2="672.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_58"
-            x1="309"
-            y1="674.25"
-            x2="299"
-            y2="674.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_59"
-            x1="309"
-            y1="676.25"
-            x2="299"
-            y2="676.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_60"
-            x1="309"
-            y1="678.25"
-            x2="299"
-            y2="678.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_61"
-            x1="309"
-            y1="682.25"
-            x2="299"
-            y2="682.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_62"
-            x1="309"
-            y1="684.25"
-            x2="299"
-            y2="684.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_63"
-            x1="309"
-            y1="686.25"
-            x2="299"
-            y2="686.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_64"
-            x1="309"
-            y1="688.25"
-            x2="299"
-            y2="688.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_65"
-            x1="309"
-            y1="692.25"
-            x2="299"
-            y2="692.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_66"
-            x1="309"
-            y1="694.25"
-            x2="299"
-            y2="694.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_67"
-            x1="309"
-            y1="696.25"
-            x2="299"
-            y2="696.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_68"
-            x1="309"
-            y1="698.25"
-            x2="299"
-            y2="698.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_69"
-            x1="309"
-            y1="702.25"
-            x2="299"
-            y2="702.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_70"
-            x1="309"
-            y1="704.25"
-            x2="299"
-            y2="704.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_71"
-            x1="309"
-            y1="706.25"
-            x2="299"
-            y2="706.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_72"
-            x1="309"
-            y1="708.25"
-            x2="299"
-            y2="708.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-        </g>
-        <g id="Group 5_2">
-          <line
-            id="Line 22_73"
-            x1="319"
-            y1="630.25"
-            x2="299"
-            y2="630.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_74"
-            x1="314"
-            y1="640.25"
-            x2="299"
-            y2="640.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_75"
-            x1="319"
-            y1="650.25"
-            x2="299"
-            y2="650.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_76"
-            x1="314"
-            y1="660.25"
-            x2="299"
-            y2="660.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_77"
-            x1="319"
-            y1="670.25"
-            x2="299"
-            y2="670.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_78"
-            x1="314"
-            y1="680.25"
-            x2="299"
-            y2="680.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_79"
-            x1="319"
-            y1="690.25"
-            x2="299"
-            y2="690.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 22_80"
-            x1="314"
-            y1="700.25"
-            x2="299"
-            y2="700.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-          <line
-            id="Line 23_2"
-            x1="319"
-            y1="710.25"
-            x2="299"
-            y2="710.25"
-            stroke="black"
-            stroke-width="0.5"
-          />
-        </g>
-        <path
-          id="1000 mL_2"
-          d="M321.986 627.636V632H321.457V628.19H321.432L320.366 628.898V628.361L321.457 627.636H321.986ZM324.664 632.06C324.343 632.06 324.07 631.972 323.844 631.798C323.618 631.621 323.445 631.366 323.326 631.033C323.207 630.697 323.147 630.293 323.147 629.818C323.147 629.347 323.207 628.944 323.326 628.61C323.447 628.275 323.62 628.019 323.846 627.843C324.073 627.665 324.346 627.577 324.664 627.577C324.982 627.577 325.254 627.665 325.48 627.843C325.707 628.019 325.881 628.275 326 628.61C326.121 628.944 326.181 629.347 326.181 629.818C326.181 630.293 326.121 630.697 326.002 631.033C325.883 631.366 325.71 631.621 325.484 631.798C325.259 631.972 324.985 632.06 324.664 632.06ZM324.664 631.591C324.982 631.591 325.229 631.437 325.406 631.131C325.582 630.824 325.67 630.386 325.67 629.818C325.67 629.44 325.629 629.119 325.548 628.853C325.469 628.587 325.354 628.385 325.203 628.246C325.054 628.107 324.874 628.037 324.664 628.037C324.349 628.037 324.102 628.192 323.925 628.504C323.747 628.813 323.658 629.251 323.658 629.818C323.658 630.196 323.698 630.517 323.778 630.781C323.857 631.045 323.972 631.246 324.121 631.384C324.271 631.522 324.452 631.591 324.664 631.591ZM328.414 632.06C328.093 632.06 327.82 631.972 327.594 631.798C327.368 631.621 327.195 631.366 327.076 631.033C326.957 630.697 326.897 630.293 326.897 629.818C326.897 629.347 326.957 628.944 327.076 628.61C327.197 628.275 327.37 628.019 327.596 627.843C327.823 627.665 328.096 627.577 328.414 627.577C328.732 627.577 329.004 627.665 329.23 627.843C329.457 628.019 329.631 628.275 329.75 628.61C329.871 628.944 329.931 629.347 329.931 629.818C329.931 630.293 329.871 630.697 329.752 631.033C329.633 631.366 329.46 631.621 329.234 631.798C329.009 631.972 328.735 632.06 328.414 632.06ZM328.414 631.591C328.732 631.591 328.979 631.437 329.156 631.131C329.332 630.824 329.42 630.386 329.42 629.818C329.42 629.44 329.379 629.119 329.298 628.853C329.219 628.587 329.104 628.385 328.953 628.246C328.804 628.107 328.624 628.037 328.414 628.037C328.099 628.037 327.852 628.192 327.675 628.504C327.497 628.813 327.408 629.251 327.408 629.818C327.408 630.196 327.448 630.517 327.528 630.781C327.607 631.045 327.722 631.246 327.871 631.384C328.021 631.522 328.202 631.591 328.414 631.591ZM332.164 632.06C331.843 632.06 331.57 631.972 331.344 631.798C331.118 631.621 330.945 631.366 330.826 631.033C330.707 630.697 330.647 630.293 330.647 629.818C330.647 629.347 330.707 628.944 330.826 628.61C330.947 628.275 331.12 628.019 331.346 627.843C331.573 627.665 331.846 627.577 332.164 627.577C332.482 627.577 332.754 627.665 332.98 627.843C333.207 628.019 333.381 628.275 333.5 628.61C333.621 628.944 333.681 629.347 333.681 629.818C333.681 630.293 333.621 630.697 333.502 631.033C333.383 631.366 333.21 631.621 332.984 631.798C332.759 631.972 332.485 632.06 332.164 632.06ZM332.164 631.591C332.482 631.591 332.729 631.437 332.906 631.131C333.082 630.824 333.17 630.386 333.17 629.818C333.17 629.44 333.129 629.119 333.048 628.853C332.969 628.587 332.854 628.385 332.703 628.246C332.554 628.107 332.374 628.037 332.164 628.037C331.849 628.037 331.602 628.192 331.425 628.504C331.247 628.813 331.158 629.251 331.158 629.818C331.158 630.196 331.198 630.517 331.278 630.781C331.357 631.045 331.472 631.246 331.621 631.384C331.771 631.522 331.952 631.591 332.164 631.591ZM336.187 632V628.727H336.673V629.239H336.715C336.783 629.064 336.893 628.928 337.045 628.832C337.197 628.734 337.38 628.685 337.593 628.685C337.809 628.685 337.989 628.734 338.132 628.832C338.277 628.928 338.39 629.064 338.471 629.239H338.505C338.589 629.07 338.714 628.935 338.882 628.836C339.05 628.735 339.251 628.685 339.485 628.685C339.778 628.685 340.017 628.776 340.203 628.96C340.389 629.141 340.482 629.425 340.482 629.81V632H339.979V629.81C339.979 629.568 339.913 629.396 339.781 629.292C339.649 629.188 339.494 629.136 339.315 629.136C339.085 629.136 338.906 629.206 338.78 629.345C338.653 629.483 338.59 629.658 338.59 629.869V632H338.079V629.759C338.079 629.572 338.018 629.423 337.898 629.309C337.777 629.194 337.621 629.136 337.431 629.136C337.3 629.136 337.178 629.171 337.065 629.241C336.952 629.31 336.862 629.407 336.792 629.531C336.724 629.653 336.69 629.794 336.69 629.955V632H336.187ZM341.47 632V627.636H341.998V631.531H344.027V632H341.47Z"
-          fill="black"
-        />
-        <path
-          id="800_2"
-          d="M323.849 652.06C323.557 652.06 323.298 652.008 323.074 651.904C322.851 651.799 322.677 651.655 322.552 651.472C322.427 651.287 322.365 651.077 322.366 650.841C322.365 650.656 322.401 650.486 322.475 650.33C322.549 650.172 322.65 650.04 322.778 649.935C322.907 649.829 323.051 649.761 323.21 649.733V649.707C323.001 649.653 322.835 649.536 322.712 649.356C322.588 649.174 322.527 648.967 322.528 648.736C322.527 648.514 322.583 648.316 322.697 648.141C322.81 647.967 322.967 647.829 323.165 647.728C323.366 647.627 323.594 647.577 323.849 647.577C324.102 647.577 324.328 647.627 324.527 647.728C324.726 647.829 324.882 647.967 324.996 648.141C325.111 648.316 325.169 648.514 325.17 648.736C325.169 648.967 325.106 649.174 324.981 649.356C324.857 649.536 324.693 649.653 324.489 649.707V649.733C324.646 649.761 324.788 649.829 324.915 649.935C325.041 650.04 325.142 650.172 325.217 650.33C325.293 650.486 325.331 650.656 325.332 650.841C325.331 651.077 325.267 651.287 325.141 651.472C325.016 651.655 324.842 651.799 324.619 651.904C324.397 652.008 324.141 652.06 323.849 652.06ZM323.849 651.591C324.047 651.591 324.217 651.559 324.361 651.495C324.504 651.431 324.615 651.341 324.693 651.224C324.771 651.108 324.811 650.972 324.812 650.815C324.811 650.651 324.768 650.505 324.685 650.379C324.601 650.252 324.487 650.153 324.342 650.08C324.198 650.008 324.034 649.972 323.849 649.972C323.663 649.972 323.497 650.008 323.351 650.08C323.206 650.153 323.092 650.252 323.008 650.379C322.925 650.505 322.885 650.651 322.886 650.815C322.885 650.972 322.923 651.108 322.999 651.224C323.077 651.341 323.189 651.431 323.334 651.495C323.479 651.559 323.651 651.591 323.849 651.591ZM323.849 649.52C324.006 649.52 324.144 649.489 324.265 649.426C324.387 649.364 324.483 649.276 324.553 649.164C324.622 649.052 324.658 648.92 324.659 648.77C324.658 648.622 324.623 648.494 324.555 648.384C324.487 648.273 324.392 648.188 324.271 648.129C324.151 648.067 324.01 648.037 323.849 648.037C323.686 648.037 323.543 648.067 323.421 648.129C323.299 648.188 323.205 648.273 323.138 648.384C323.071 648.494 323.038 648.622 323.04 648.77C323.038 648.92 323.072 649.052 323.14 649.164C323.21 649.276 323.305 649.364 323.428 649.426C323.55 649.489 323.69 649.52 323.849 649.52ZM327.572 652.06C327.251 652.06 326.978 651.972 326.752 651.798C326.526 651.621 326.354 651.366 326.234 651.033C326.115 650.697 326.055 650.293 326.055 649.818C326.055 649.347 326.115 648.944 326.234 648.61C326.355 648.275 326.528 648.019 326.754 647.843C326.981 647.665 327.254 647.577 327.572 647.577C327.89 647.577 328.162 647.665 328.388 647.843C328.616 648.019 328.789 648.275 328.908 648.61C329.029 648.944 329.089 649.347 329.089 649.818C329.089 650.293 329.03 650.697 328.91 651.033C328.791 651.366 328.618 651.621 328.393 651.798C328.167 651.972 327.893 652.06 327.572 652.06ZM327.572 651.591C327.89 651.591 328.138 651.437 328.314 651.131C328.49 650.824 328.578 650.386 328.578 649.818C328.578 649.44 328.537 649.119 328.456 648.853C328.377 648.587 328.262 648.385 328.111 648.246C327.962 648.107 327.782 648.037 327.572 648.037C327.257 648.037 327.01 648.192 326.833 648.504C326.655 648.813 326.567 649.251 326.567 649.818C326.567 650.196 326.606 650.517 326.686 650.781C326.765 651.045 326.88 651.246 327.029 651.384C327.18 651.522 327.361 651.591 327.572 651.591ZM331.322 652.06C331.001 652.06 330.728 651.972 330.502 651.798C330.276 651.621 330.104 651.366 329.984 651.033C329.865 650.697 329.805 650.293 329.805 649.818C329.805 649.347 329.865 648.944 329.984 648.61C330.105 648.275 330.278 648.019 330.504 647.843C330.731 647.665 331.004 647.577 331.322 647.577C331.64 647.577 331.912 647.665 332.138 647.843C332.366 648.019 332.539 648.275 332.658 648.61C332.779 648.944 332.839 649.347 332.839 649.818C332.839 650.293 332.78 650.697 332.66 651.033C332.541 651.366 332.368 651.621 332.143 651.798C331.917 651.972 331.643 652.06 331.322 652.06ZM331.322 651.591C331.64 651.591 331.888 651.437 332.064 651.131C332.24 650.824 332.328 650.386 332.328 649.818C332.328 649.44 332.287 649.119 332.206 648.853C332.127 648.587 332.012 648.385 331.861 648.246C331.712 648.107 331.532 648.037 331.322 648.037C331.007 648.037 330.76 648.192 330.583 648.504C330.405 648.813 330.317 649.251 330.317 649.818C330.317 650.196 330.356 650.517 330.436 650.781C330.515 651.045 330.63 651.246 330.779 651.384C330.93 651.522 331.111 651.591 331.322 651.591Z"
-          fill="black"
-        />
-        <path
-          id="600_2"
-          d="M323.901 672.06C323.722 672.057 323.543 672.023 323.364 671.957C323.185 671.892 323.021 671.782 322.874 671.627C322.726 671.471 322.607 671.26 322.518 670.994C322.428 670.727 322.384 670.392 322.384 669.989C322.384 669.602 322.42 669.26 322.492 668.962C322.565 668.662 322.67 668.41 322.808 668.205C322.945 667.999 323.112 667.843 323.306 667.737C323.502 667.63 323.723 667.577 323.969 667.577C324.213 667.577 324.43 667.626 324.621 667.724C324.813 667.82 324.969 667.955 325.089 668.129C325.21 668.302 325.288 668.501 325.324 668.727H324.804C324.756 668.531 324.662 668.369 324.523 668.239C324.384 668.11 324.199 668.045 323.969 668.045C323.631 668.045 323.364 668.192 323.17 668.487C322.977 668.781 322.879 669.193 322.878 669.724H322.912C322.991 669.604 323.086 669.501 323.195 669.415C323.306 669.329 323.428 669.262 323.562 669.215C323.695 669.168 323.837 669.145 323.986 669.145C324.236 669.145 324.464 669.207 324.672 669.332C324.879 669.456 325.045 669.627 325.17 669.846C325.295 670.063 325.358 670.312 325.358 670.594C325.358 670.864 325.298 671.111 325.177 671.335C325.056 671.558 324.886 671.736 324.668 671.868C324.45 671.999 324.195 672.062 323.901 672.06ZM323.901 671.591C324.08 671.591 324.24 671.546 324.382 671.457C324.526 671.367 324.638 671.247 324.721 671.097C324.805 670.946 324.847 670.778 324.847 670.594C324.847 670.413 324.806 670.249 324.725 670.102C324.646 669.952 324.536 669.834 324.395 669.746C324.256 669.658 324.097 669.614 323.918 669.614C323.783 669.614 323.657 669.641 323.54 669.695C323.424 669.747 323.322 669.82 323.234 669.912C323.147 670.004 323.079 670.11 323.029 670.229C322.979 670.347 322.955 670.472 322.955 670.602C322.955 670.776 322.995 670.937 323.076 671.088C323.158 671.239 323.271 671.36 323.413 671.452C323.556 671.545 323.719 671.591 323.901 671.591ZM327.619 672.06C327.298 672.06 327.025 671.972 326.799 671.798C326.573 671.621 326.4 671.366 326.281 671.033C326.162 670.697 326.102 670.293 326.102 669.818C326.102 669.347 326.162 668.944 326.281 668.61C326.402 668.275 326.575 668.019 326.801 667.843C327.028 667.665 327.301 667.577 327.619 667.577C327.937 667.577 328.209 667.665 328.435 667.843C328.662 668.019 328.836 668.275 328.955 668.61C329.076 668.944 329.136 669.347 329.136 669.818C329.136 670.293 329.077 670.697 328.957 671.033C328.838 671.366 328.665 671.621 328.439 671.798C328.214 671.972 327.94 672.06 327.619 672.06ZM327.619 671.591C327.937 671.591 328.184 671.437 328.361 671.131C328.537 670.824 328.625 670.386 328.625 669.818C328.625 669.44 328.584 669.119 328.503 668.853C328.424 668.587 328.309 668.385 328.158 668.246C328.009 668.107 327.829 668.037 327.619 668.037C327.304 668.037 327.057 668.192 326.88 668.504C326.702 668.813 326.613 669.251 326.613 669.818C326.613 670.196 326.653 670.517 326.733 670.781C326.812 671.045 326.927 671.246 327.076 671.384C327.226 671.522 327.407 671.591 327.619 671.591ZM331.369 672.06C331.048 672.06 330.775 671.972 330.549 671.798C330.323 671.621 330.15 671.366 330.031 671.033C329.912 670.697 329.852 670.293 329.852 669.818C329.852 669.347 329.912 668.944 330.031 668.61C330.152 668.275 330.325 668.019 330.551 667.843C330.778 667.665 331.051 667.577 331.369 667.577C331.687 667.577 331.959 667.665 332.185 667.843C332.412 668.019 332.586 668.275 332.705 668.61C332.826 668.944 332.886 669.347 332.886 669.818C332.886 670.293 332.827 670.697 332.707 671.033C332.588 671.366 332.415 671.621 332.189 671.798C331.964 671.972 331.69 672.06 331.369 672.06ZM331.369 671.591C331.687 671.591 331.934 671.437 332.111 671.131C332.287 670.824 332.375 670.386 332.375 669.818C332.375 669.44 332.334 669.119 332.253 668.853C332.174 668.587 332.059 668.385 331.908 668.246C331.759 668.107 331.579 668.037 331.369 668.037C331.054 668.037 330.807 668.192 330.63 668.504C330.452 668.813 330.363 669.251 330.363 669.818C330.363 670.196 330.403 670.517 330.483 670.781C330.562 671.045 330.677 671.246 330.826 671.384C330.976 671.522 331.157 671.591 331.369 671.591Z"
-          fill="black"
-        />
-        <path
-          id="400_2"
-          d="M322.349 691.105V690.67L324.267 687.636H324.582V688.31H324.369L322.92 690.602V690.636H325.503V691.105H322.349ZM324.403 692V690.973V690.771V687.636H324.906V692H324.403ZM327.725 692.06C327.404 692.06 327.13 691.972 326.904 691.798C326.678 691.621 326.506 691.366 326.387 691.033C326.267 690.697 326.208 690.293 326.208 689.818C326.208 689.347 326.267 688.944 326.387 688.61C326.507 688.275 326.681 688.019 326.906 687.843C327.134 687.665 327.406 687.577 327.725 687.577C328.043 687.577 328.315 687.665 328.541 687.843C328.768 688.019 328.941 688.275 329.061 688.61C329.181 688.944 329.242 689.347 329.242 689.818C329.242 690.293 329.182 690.697 329.063 691.033C328.943 691.366 328.771 691.621 328.545 691.798C328.319 691.972 328.046 692.06 327.725 692.06ZM327.725 691.591C328.043 691.591 328.29 691.437 328.466 691.131C328.642 690.824 328.73 690.386 328.73 689.818C328.73 689.44 328.69 689.119 328.609 688.853C328.529 688.587 328.414 688.385 328.264 688.246C328.115 688.107 327.935 688.037 327.725 688.037C327.409 688.037 327.163 688.192 326.985 688.504C326.808 688.813 326.719 689.251 326.719 689.818C326.719 690.196 326.759 690.517 326.838 690.781C326.918 691.045 327.032 691.246 327.181 691.384C327.332 691.522 327.513 691.591 327.725 691.591ZM331.475 692.06C331.154 692.06 330.88 691.972 330.654 691.798C330.428 691.621 330.256 691.366 330.137 691.033C330.017 690.697 329.958 690.293 329.958 689.818C329.958 689.347 330.017 688.944 330.137 688.61C330.257 688.275 330.431 688.019 330.656 687.843C330.884 687.665 331.156 687.577 331.475 687.577C331.793 687.577 332.065 687.665 332.291 687.843C332.518 688.019 332.691 688.275 332.811 688.61C332.931 688.944 332.992 689.347 332.992 689.818C332.992 690.293 332.932 690.697 332.813 691.033C332.693 691.366 332.521 691.621 332.295 691.798C332.069 691.972 331.796 692.06 331.475 692.06ZM331.475 691.591C331.793 691.591 332.04 691.437 332.216 691.131C332.392 690.824 332.48 690.386 332.48 689.818C332.48 689.44 332.44 689.119 332.359 688.853C332.279 688.587 332.164 688.385 332.014 688.246C331.865 688.107 331.685 688.037 331.475 688.037C331.159 688.037 330.913 688.192 330.735 688.504C330.558 688.813 330.469 689.251 330.469 689.818C330.469 690.196 330.509 690.517 330.588 690.781C330.668 691.045 330.782 691.246 330.931 691.384C331.082 691.522 331.263 691.591 331.475 691.591Z"
-          fill="black"
-        />
-        <path
-          id="200_2"
-          d="M322.452 712V711.616L323.892 710.04C324.061 709.855 324.2 709.695 324.31 709.558C324.419 709.42 324.5 709.291 324.553 709.17C324.607 709.048 324.634 708.92 324.634 708.787C324.634 708.634 324.597 708.501 324.523 708.388C324.45 708.276 324.351 708.19 324.224 708.129C324.098 708.067 323.956 708.037 323.798 708.037C323.631 708.037 323.484 708.072 323.359 708.141C323.236 708.21 323.14 708.305 323.072 708.429C323.005 708.553 322.972 708.697 322.972 708.864H322.469C322.469 708.608 322.528 708.384 322.646 708.19C322.763 707.997 322.924 707.847 323.127 707.739C323.332 707.631 323.561 707.577 323.815 707.577C324.071 707.577 324.298 707.631 324.495 707.739C324.692 707.847 324.847 707.992 324.96 708.175C325.072 708.359 325.128 708.562 325.128 708.787C325.128 708.947 325.099 709.104 325.04 709.258C324.984 709.41 324.884 709.58 324.742 709.767C324.602 709.953 324.406 710.18 324.156 710.449L323.176 711.497V711.531H325.205V712H322.452ZM327.508 712.06C327.187 712.06 326.913 711.972 326.688 711.798C326.462 711.621 326.289 711.366 326.17 711.033C326.05 710.697 325.991 710.293 325.991 709.818C325.991 709.347 326.05 708.944 326.17 708.61C326.29 708.275 326.464 708.019 326.69 707.843C326.917 707.665 327.19 707.577 327.508 707.577C327.826 707.577 328.098 707.665 328.324 707.843C328.551 708.019 328.724 708.275 328.844 708.61C328.964 708.944 329.025 709.347 329.025 709.818C329.025 710.293 328.965 710.697 328.846 711.033C328.727 711.366 328.554 711.621 328.328 711.798C328.102 711.972 327.829 712.06 327.508 712.06ZM327.508 711.591C327.826 711.591 328.073 711.437 328.249 711.131C328.425 710.824 328.513 710.386 328.513 709.818C328.513 709.44 328.473 709.119 328.392 708.853C328.313 708.587 328.197 708.385 328.047 708.246C327.898 708.107 327.718 708.037 327.508 708.037C327.192 708.037 326.946 708.192 326.768 708.504C326.591 708.813 326.502 709.251 326.502 709.818C326.502 710.196 326.542 710.517 326.621 710.781C326.701 711.045 326.815 711.246 326.964 711.384C327.115 711.522 327.296 711.591 327.508 711.591ZM331.258 712.06C330.937 712.06 330.663 711.972 330.438 711.798C330.212 711.621 330.039 711.366 329.92 711.033C329.8 710.697 329.741 710.293 329.741 709.818C329.741 709.347 329.8 708.944 329.92 708.61C330.04 708.275 330.214 708.019 330.44 707.843C330.667 707.665 330.94 707.577 331.258 707.577C331.576 707.577 331.848 707.665 332.074 707.843C332.301 708.019 332.474 708.275 332.594 708.61C332.714 708.944 332.775 709.347 332.775 709.818C332.775 710.293 332.715 710.697 332.596 711.033C332.477 711.366 332.304 711.621 332.078 711.798C331.852 711.972 331.579 712.06 331.258 712.06ZM331.258 711.591C331.576 711.591 331.823 711.437 331.999 711.131C332.175 710.824 332.263 710.386 332.263 709.818C332.263 709.44 332.223 709.119 332.142 708.853C332.063 708.587 331.947 708.385 331.797 708.246C331.648 708.107 331.468 708.037 331.258 708.037C330.942 708.037 330.696 708.192 330.518 708.504C330.341 708.813 330.252 709.251 330.252 709.818C330.252 710.196 330.292 710.517 330.371 710.781C330.451 711.045 330.565 711.246 330.714 711.384C330.865 711.522 331.046 711.591 331.258 711.591Z"
-          fill="black"
-        />
-        <line
-          id="Line 24_2"
-          x1="299.25"
-          y1="630"
-          x2="299.25"
-          y2="710"
-          stroke="black"
-          stroke-width="0.5"
-        />
-      </g>
-    </g>
-  </g>
-  <StaticDefs/>
-</svg>);
-}
+)}
 
-export default Apparatus;
+export default Rotameter;
