+++ lab-experiments/kettle-boiler/src/components/Interactables/Waterfall.tsx
@@ -1,0 +1,8 @@
+import type { Component } from "solid-js";
+
+export const Waterfall: Component = () => {
+  return (
+    <></>
+)};
+
+export default Waterfall;
