+++ lab-experiments/kettle-boiler/src/components/Static/StaticDefs.tsx
