+++ lab-experiments/kettle-boiler/src/components/Static/StaticElements.tsx
